Methods for `AdjacencyMatrix` objects

`length` returns the length of an `AdjacencyMatrix` object (number of rows of an assay). `length` returns a `numeric` of length 1.

`dim` returns the length of an `AdjacencyMatrix` object (number of rows of an assay, number of cols of an assay). `dim` returns a `numeric` of length 2.

`type` will return the type of an `AdjacencyMatrix` (`statistical`, `structural` or `combine`). `type` returns a `character` of length 1

`directed` returns the information on directed of an `AdjacencyMatrix`, i.e. if the underlying graph is directed or undirected. `directed` returns `logical` of length 1.

`thresholded` returns the information if the adjacency matrix is thresholded, i.e. if the function `rtCorrection` or `threshold` was applied to the `AdjacencyMatrix` object. `thresholded` returns a `logical` of length 1.

`show` prints summary information on an object of class `AdjacencyMatrix`.

`as.data.frame` returns the adjacency matrices (stored in the `assays` slot) and returns information on the nodes and the associated information on edges as a data frame. `as.data.frame` returns a `data.frame`.

# S4 method for class 'AdjacencyMatrix'
length(x)

# S4 method for class 'AdjacencyMatrix'
dim(x)

# S4 method for class 'AdjacencyMatrix'
type(x)

# S4 method for class 'AdjacencyMatrix'
directed(object)

# S4 method for class 'AdjacencyMatrix'
thresholded(object)

# S4 method for class 'AdjacencyMatrix'
show(object)

# S4 method for class 'AdjacencyMatrix'
as.data.frame(x)

Arguments

x

instance of class `AdjacencyMatrix`

object

instance of class `AdjacencyMatrix`

Author

Thomas Naake, thomasnaake@googlemail.com