Create an adjacency matrix based on algorithm for the reconstruction of accurate cellular networks

`aracne` infers an adjacency matrix using the algorithm for the reconstruction of accurate cellular networks using the `aracne.a` function from the `parmigene` package. The function `aracne` will return the weighted adjacency matrix of the inferred network after applying `aracne.a`.

aracne(mi, eps = 0.05, ...)

Arguments

mi

matrix, where columns are the samples and the rows are features (metabolites), cell entries are mutual information values between the features. As input, the mutual information (e.g. raw MI estimates) from the `knnmi.all` function of the `parmigene` package can be used.

eps

numeric, used to remove the weakest edge of each triple of nodes

...

not used here

Value

matrix, matrix with edges inferred from Reconstruction of accurate cellular networks algorithm `aracne.a`

Details

For more details on the `aracne.a` function, refer to `?parmigene::aracne.a`. `aracne.a` considers each triple of edges independently and removes the weakest one if \(MI(i, j) < MI(j, k) - eps\) and \(MI(i, j) < MI(i, k) - eps\). See Margolin et al. (2006) for further information.

References

Margolin et al. (2006): ARACNE : An algorithm for the reconstruction of gene regulatory networks in a mammalian cellular context. BMC Bioinformatics, S7, doi: [10.1186/1471-2105-7-S1-S7](https://doi.org/10.1186/1471-2105-7-S1-S7)

Author

Thomas Naake, thomasnaake@googlemail.com

Examples

data("x_test", package = "MetNet")
x <- x_test[1:10, 3:ncol(x_test)]
x <- as.matrix(x)
x_z <- apply(x, 1, function(y) (y - mean(y)) / sd(y))
mi_x_z <- parmigene::knnmi.all(x_z)
aracne(mi_x_z, eps = 0.05)
#>                   X001_BA4_01_28001 X002_BA5_01_28002 X003_BA6_01_28003
#> X001_BA4_01_28001         0.0000000         0.0000000         0.0000000
#> X002_BA5_01_28002         0.0000000         0.0000000         0.6123016
#> X003_BA6_01_28003         0.0000000         0.6123016         0.0000000
#> X004_BA7_01_28004         0.0000000         0.0000000         0.0000000
#> X005_BA8_01_28005         0.0000000         0.0000000         0.0000000
#> X011_BB6_01_28012         0.0000000         0.0000000         0.7123016
#> X012_BB7_01_28013         0.0000000         0.0000000         0.1730159
#> X013_BB8_01_28014         0.0000000         0.0000000         0.0000000
#>                   X004_BA7_01_28004 X005_BA8_01_28005 X011_BB6_01_28012
#> X001_BA4_01_28001         0.0000000         0.0000000         0.0000000
#> X002_BA5_01_28002         0.0000000         0.0000000         0.0000000
#> X003_BA6_01_28003         0.0000000         0.0000000         0.7123016
#> X004_BA7_01_28004         0.0000000         0.0000000         0.0000000
#> X005_BA8_01_28005         0.0000000         0.0000000         0.0000000
#> X011_BB6_01_28012         0.0000000         0.0000000         0.0000000
#> X012_BB7_01_28013         0.0000000         0.0000000         0.2170635
#> X013_BB8_01_28014         0.0000000         0.0000000         0.0000000
#>                   X012_BB7_01_28013 X013_BB8_01_28014 X014_BC1_01_28015
#> X001_BA4_01_28001         0.0000000         0.0000000         0.0000000
#> X002_BA5_01_28002         0.0000000         0.0000000         0.3489683
#> X003_BA6_01_28003         0.1730159         0.0000000         0.0000000
#> X004_BA7_01_28004         0.0000000         0.0000000         0.4039683
#> X005_BA8_01_28005         0.0000000         0.0000000         0.0000000
#> X011_BB6_01_28012         0.2170635         0.0000000         0.0000000
#> X012_BB7_01_28013         0.0000000         0.0000000         0.6223016
#> X013_BB8_01_28014         0.0000000         0.0000000         0.5589683
#>                   X015_BC2_01_28016 X016_BC3_01_28017 X017_BC4_01_28018
#> X001_BA4_01_28001         0.0000000         0.0000000         0.0000000
#> X002_BA5_01_28002         0.0000000         0.0000000         0.5905159
#> X003_BA6_01_28003         0.0000000         0.0000000         0.0000000
#> X004_BA7_01_28004         0.0000000         0.0000000         0.0000000
#> X005_BA8_01_28005         0.0000000         0.0000000         0.0000000
#> X011_BB6_01_28012         0.6589683         0.0000000         0.6821825
#> X012_BB7_01_28013         0.0000000         0.0000000         0.0000000
#> X013_BB8_01_28014         0.0000000         0.0000000         0.0000000
#>                   X018_BC5_01_28019 X019_BC6_01_28020 X020_BC7_01_28021
#> X001_BA4_01_28001         0.0000000         0.0000000        0.00000000
#> X002_BA5_01_28002         0.0000000         0.0000000        0.00000000
#> X003_BA6_01_28003         0.0000000         0.0000000        0.00000000
#> X004_BA7_01_28004         0.0000000         0.0000000        0.00000000
#> X005_BA8_01_28005         0.0000000         0.0000000        0.00000000
#> X011_BB6_01_28012         0.0000000         0.0000000        0.00000000
#> X012_BB7_01_28013         0.0000000         0.0000000        0.00000000
#> X013_BB8_01_28014         0.0000000         0.0000000        0.00000000
#>                   X026_BD5_01_28028 X027_BD6_01_28029 X028_BD7_01_28030
#> X001_BA4_01_28001         0.0000000         0.0000000         0.0000000
#> X002_BA5_01_28002         0.0000000         0.0000000         0.0000000
#> X003_BA6_01_28003         0.0000000         0.0000000         0.0000000
#> X004_BA7_01_28004         0.0000000         0.5056349         0.0000000
#> X005_BA8_01_28005         0.0000000         0.0000000         0.7206349
#> X011_BB6_01_28012         0.5030159         0.0000000         0.0000000
#> X012_BB7_01_28013         0.0000000         0.0000000         0.0000000
#> X013_BB8_01_28014         0.0000000         0.0000000         0.0000000
#>                   X029_BD8_01_28031 X030_BE1_01_28032 X031_BE2_01_28034
#> X001_BA4_01_28001         0.0000000         0.0000000         0.0000000
#> X002_BA5_01_28002         0.0000000         0.0000000         0.0000000
#> X003_BA6_01_28003         0.0000000         0.0000000         0.0000000
#> X004_BA7_01_28004         0.0000000         0.0000000         0.0000000
#> X005_BA8_01_28005         0.0000000         0.0000000         0.0000000
#> X011_BB6_01_28012         0.4830159         0.0000000         0.0000000
#> X012_BB7_01_28013         0.0000000         0.0000000         0.0000000
#> X013_BB8_01_28014         0.0000000         0.0000000         0.0000000
#>                   X032_BE3_01_28035 X033_BE4_01_28036 X034_BE5_01_28037
#> X001_BA4_01_28001         0.0000000         0.2323016         0.0000000
#> X002_BA5_01_28002         0.0000000         0.0000000         0.4580159
#> X003_BA6_01_28003         0.0000000         0.0000000         0.0000000
#> X004_BA7_01_28004         0.0000000         0.0000000         0.0000000
#> X005_BA8_01_28005         0.0000000         0.0000000         0.0000000
#> X011_BB6_01_28012         0.0000000         0.0000000         0.0000000
#> X012_BB7_01_28013         0.0000000         0.0000000         0.0000000
#> X013_BB8_01_28014         0.0000000         0.0000000         0.0000000
#>                   X035_BE6_01_28038 X041_RA4_01_28045 X042_RA5_01_28046
#> X001_BA4_01_28001         0.0000000         0.0000000         0.0000000
#> X002_BA5_01_28002         0.1930159         0.0000000         0.0000000
#> X003_BA6_01_28003         0.0000000         0.0000000         0.0000000
#> X004_BA7_01_28004         0.0000000         0.0000000         0.0000000
#> X005_BA8_01_28005         0.0000000         0.0000000         0.0000000
#> X011_BB6_01_28012         0.0000000         0.0000000         0.0000000
#> X012_BB7_01_28013         0.0000000         0.0000000         0.0000000
#> X013_BB8_01_28014         0.0000000         0.0000000         0.0000000
#>                   X043_RA6_01_28047 X044_RA7_01_28048 X045_RA8_01_28049
#> X001_BA4_01_28001         0.0000000         0.0000000         0.0000000
#> X002_BA5_01_28002         0.0000000         0.0000000         0.0000000
#> X003_BA6_01_28003         0.0000000         0.0000000         0.0000000
#> X004_BA7_01_28004         0.0000000         0.0000000         0.2830159
#> X005_BA8_01_28005         0.0000000         0.0000000         0.0000000
#> X011_BB6_01_28012         0.0000000         0.0000000         0.0000000
#> X012_BB7_01_28013         0.4996825         0.0000000         0.0000000
#> X013_BB8_01_28014         0.0000000         0.0000000         0.0000000
#>                   X046_RB1_01_28050 X047_RB2_01_28051 X048_RB3_01_28052
#> X001_BA4_01_28001         0.0000000         0.0000000         0.0000000
#> X002_BA5_01_28002         0.0000000         0.0000000         0.0000000
#> X003_BA6_01_28003         0.0000000         0.0000000         0.0000000
#> X004_BA7_01_28004         0.0000000         0.0000000         0.0000000
#> X005_BA8_01_28005         0.0000000         0.0000000         0.0000000
#> X011_BB6_01_28012         0.0000000         0.0000000         0.0000000
#> X012_BB7_01_28013         0.0000000         0.0000000         0.0000000
#> X013_BB8_01_28014         0.0000000         0.0000000         0.0000000
#>                   X049_RB4_01_28053 X050_RB5_01_28054 X056_RC3_01_28061
#> X001_BA4_01_28001         0.0000000         0.0000000         0.0000000
#> X002_BA5_01_28002         0.0000000         0.0000000         0.0000000
#> X003_BA6_01_28003         0.0000000         0.0000000         0.0000000
#> X004_BA7_01_28004         0.0000000         0.0000000         0.0000000
#> X005_BA8_01_28005         0.0000000         0.0000000         0.0000000
#> X011_BB6_01_28012         0.0000000         0.0000000         0.0000000
#> X012_BB7_01_28013         0.0000000         0.0000000         0.0000000
#> X013_BB8_01_28014         0.0000000         0.0000000         0.0000000
#>                   X057_RC4_01_28062 X058_RC5_01_28063 X059_RC6_01_28064
#> X001_BA4_01_28001         0.0000000         0.0000000         0.0000000
#> X002_BA5_01_28002         0.0000000         0.0000000         0.0000000
#> X003_BA6_01_28003         0.0000000         0.0000000         0.0000000
#> X004_BA7_01_28004         0.0000000         0.0000000         0.0000000
#> X005_BA8_01_28005         0.0000000         0.0000000         0.0000000
#> X011_BB6_01_28012         0.0000000         0.0000000         0.0000000
#> X012_BB7_01_28013         0.0000000         0.0000000         0.0000000
#> X013_BB8_01_28014         0.0000000         0.0000000         0.0000000
#>                   X060_RC7_01_28065 X061_RC8_01_28067 X062_RD1_01_28068
#> X001_BA4_01_28001         0.0000000         0.0000000         0.0000000
#> X002_BA5_01_28002         0.0000000         0.0000000         0.0000000
#> X003_BA6_01_28003         0.0000000         0.0000000         0.0000000
#> X004_BA7_01_28004         0.0000000         0.0000000         0.0000000
#> X005_BA8_01_28005         0.0000000         0.0000000         0.0000000
#> X011_BB6_01_28012         0.0000000         0.0000000         0.0000000
#> X012_BB7_01_28013         0.0000000         0.0000000         0.0000000
#> X013_BB8_01_28014         0.0000000         0.0000000         0.0000000
#>                   X063_RD2_01_28069 X064_RD3_01_28070 X065_RD4_01_28071
#> X001_BA4_01_28001         0.0000000        0.00000000         0.0000000
#> X002_BA5_01_28002         0.0000000        0.00000000         0.0000000
#> X003_BA6_01_28003         0.0000000        0.00000000         0.0000000
#> X004_BA7_01_28004         0.0000000        0.00000000         0.0000000
#> X005_BA8_01_28005         0.0000000        0.00000000         0.3706349
#> X011_BB6_01_28012         0.0000000        0.00000000         0.0000000
#> X012_BB7_01_28013         0.5173016        0.00000000         0.0000000
#> X013_BB8_01_28014         0.0000000        0.00000000         0.0000000
#>                   X071_RE2_01_28078 X072_RE3_01_28079 X073_RE4_01_28080
#> X001_BA4_01_28001         0.0000000         0.0000000         0.0000000
#> X002_BA5_01_28002         0.0000000         0.0000000         0.0000000
#> X003_BA6_01_28003         0.0000000         0.0000000         0.0000000
#> X004_BA7_01_28004         0.0000000         0.0000000         0.0000000
#> X005_BA8_01_28005         0.0000000         0.0000000         0.0000000
#> X011_BB6_01_28012         0.0000000         0.0000000         0.0000000
#> X012_BB7_01_28013         0.0000000         0.0000000         0.0000000
#> X013_BB8_01_28014         0.0000000         0.0000000         0.0000000
#>                   X074_RE5_01_28081 X075_RE6_01_28082 X076_RE7_01_28083
#> X001_BA4_01_28001         0.0000000         0.0000000         0.0000000
#> X002_BA5_01_28002         0.0000000         0.0000000         0.0000000
#> X003_BA6_01_28003         0.0000000         0.0000000         0.0000000
#> X004_BA7_01_28004         0.0000000         0.0000000         0.0000000
#> X005_BA8_01_28005         0.0000000         0.0000000         0.0000000
#> X011_BB6_01_28012         0.0000000         0.0000000         0.0000000
#> X012_BB7_01_28013         0.0000000         0.0000000         0.0000000
#> X013_BB8_01_28014         0.0000000         0.0000000         0.0000000
#>                   X077_RE8_01_28084 X078_GA1_01_28085 X079_GA2_01_28086
#> X001_BA4_01_28001         0.0000000         0.0000000         0.0000000
#> X002_BA5_01_28002         0.0000000         0.3856349         0.0000000
#> X003_BA6_01_28003         0.0000000         0.3556349         0.0000000
#> X004_BA7_01_28004         0.0000000         0.0000000         0.0000000
#> X005_BA8_01_28005         0.0000000         0.0000000         0.0000000
#> X011_BB6_01_28012         0.0000000         0.0000000         0.0000000
#> X012_BB7_01_28013         0.0000000         0.0000000         0.0000000
#> X013_BB8_01_28014         0.0000000         0.0000000         0.0000000
#>                   X080_GA3_01_28087 X086_GB1_01_28094 X087_GB2_01_28095
#> X001_BA4_01_28001         0.0000000         0.0000000         0.0000000
#> X002_BA5_01_28002         0.0000000         0.0000000         0.0000000
#> X003_BA6_01_28003         0.0000000         0.0000000         0.0000000
#> X004_BA7_01_28004         0.0000000         0.0000000         0.0000000
#> X005_BA8_01_28005         0.0000000         0.0000000         0.0000000
#> X011_BB6_01_28012         0.0000000         0.0000000         0.0000000
#> X012_BB7_01_28013         0.0000000         0.0000000         0.0000000
#> X013_BB8_01_28014         0.0000000         0.0000000         0.0000000
#>                   X088_GB3_01_28096 X089_GB4_01_28097 X090_GB5_01_28098
#> X001_BA4_01_28001         0.0000000         0.0000000         0.0000000
#> X002_BA5_01_28002         0.0000000         0.0000000         0.0000000
#> X003_BA6_01_28003         0.0000000         0.0000000         0.0000000
#> X004_BA7_01_28004         0.2723016         0.0000000         0.0000000
#> X005_BA8_01_28005         0.0000000         0.0000000         0.0000000
#> X011_BB6_01_28012         0.0000000         0.0000000         0.0000000
#> X012_BB7_01_28013         0.0000000         0.0000000         0.0000000
#> X013_BB8_01_28014         0.0000000         0.4506349         0.0000000
#>                   X091_GB6_01_28100 X092_GB7_01_28101 X093_GB8_01_28102
#> X001_BA4_01_28001         0.7173016         0.0000000         0.5789683
#> X002_BA5_01_28002         0.5673016         0.7956349         0.0000000
#> X003_BA6_01_28003         0.0000000         0.5839683         0.0000000
#> X004_BA7_01_28004         0.0000000         0.0000000         0.0000000
#> X005_BA8_01_28005         0.0000000         0.5539683         0.0000000
#> X011_BB6_01_28012         0.0000000         0.6206349         0.0000000
#> X012_BB7_01_28013         0.0000000         0.0000000         0.0000000
#> X013_BB8_01_28014         0.0000000         0.0000000         0.0000000
#>                   X094_GC1_01_28103 X095_GC2_01_28104 X101_GC8_01_28111
#> X001_BA4_01_28001         0.5706349         0.0000000         0.0000000
#> X002_BA5_01_28002         0.6506349         0.7373016         0.0000000
#> X003_BA6_01_28003         0.6256349         0.0000000         0.0000000
#> X004_BA7_01_28004         0.0000000         0.0000000         0.0000000
#> X005_BA8_01_28005         0.0000000         0.0000000         0.0000000
#> X011_BB6_01_28012         0.0000000         0.6113492         0.0000000
#> X012_BB7_01_28013         0.0000000         0.0000000         0.0000000
#> X013_BB8_01_28014         0.0000000         0.0000000         0.3037302
#>                   X102_GD1_01_28112 X103_GD2_01_28113 X104_GD3_01_28114
#> X001_BA4_01_28001         0.0000000         0.0000000         0.0000000
#> X002_BA5_01_28002         0.0000000         0.0000000         0.0000000
#> X003_BA6_01_28003         0.0000000         0.0000000         0.0000000
#> X004_BA7_01_28004         0.0000000         0.0000000         0.0000000
#> X005_BA8_01_28005         0.0000000         0.0000000         0.0000000
#> X011_BB6_01_28012         0.0000000         0.0000000         0.0000000
#> X012_BB7_01_28013         0.0000000         0.0000000         0.0000000
#> X013_BB8_01_28014         0.0000000         0.0000000         0.0000000
#>                   X105_GD4_01_28115 X106_GD5_01_28116 X107_GD6_01_28117
#> X001_BA4_01_28001         0.0000000         0.0000000         0.0000000
#> X002_BA5_01_28002         0.1723016         0.0000000         0.0000000
#> X003_BA6_01_28003         0.0000000         0.0000000         0.0000000
#> X004_BA7_01_28004         0.0000000         0.0000000         0.0000000
#> X005_BA8_01_28005         0.0000000         0.0000000         0.0000000
#> X011_BB6_01_28012         0.0000000         0.0000000         0.0000000
#> X012_BB7_01_28013         0.0000000         0.0000000         0.2830159
#> X013_BB8_01_28014         0.0000000         0.0000000         0.0000000
#>                   X108_GD7_01_28118 X109_GD8_01_28119 X110_GE1_01_28120
#> X001_BA4_01_28001         0.0000000         0.0000000         0.0000000
#> X002_BA5_01_28002         0.0000000         0.0000000         0.0000000
#> X003_BA6_01_28003         0.0000000         0.0000000         0.0000000
#> X004_BA7_01_28004         0.0000000         0.0000000         0.0000000
#> X005_BA8_01_28005         0.0000000         0.0000000         0.0000000
#> X011_BB6_01_28012         0.0000000         0.0000000         0.0000000
#> X012_BB7_01_28013         0.0000000         0.3656349         0.3796825
#> X013_BB8_01_28014         0.0000000         0.0000000         0.0000000
#>                   X116_GE7_01_28127 X117_GE8_01_28128 X118_BA4_01_28132
#> X001_BA4_01_28001         0.0000000         0.0000000         0.0000000
#> X002_BA5_01_28002         0.0000000         0.0000000         0.0000000
#> X003_BA6_01_28003         0.0000000         0.0000000         0.0000000
#> X004_BA7_01_28004         0.0000000         0.0000000         0.0000000
#> X005_BA8_01_28005         0.0000000         0.0000000         0.0000000
#> X011_BB6_01_28012         0.0000000         0.0000000         0.0000000
#> X012_BB7_01_28013         0.0000000         0.0000000         0.0000000
#> X013_BB8_01_28014         0.5756349         0.0000000         0.0000000
#>                   X119_BA5_01_28133 X120_BA6_01_28134 X121_BA7_01_28136
#> X001_BA4_01_28001         0.0000000         0.0000000         0.0000000
#> X002_BA5_01_28002         0.0000000         0.0000000         0.0000000
#> X003_BA6_01_28003         0.0000000         0.0000000         0.0000000
#> X004_BA7_01_28004         0.0000000         0.0000000         0.0000000
#> X005_BA8_01_28005         0.0000000         0.0000000         0.0000000
#> X011_BB6_01_28012         0.0000000         0.0000000         0.0000000
#> X012_BB7_01_28013         0.0000000         0.0000000         0.0000000
#> X013_BB8_01_28014         0.0000000         0.0000000         0.0000000
#>                   X122_BA8_01_28137 X123_BB1_01_28138 X124_BB2_01_28139
#> X001_BA4_01_28001         0.0000000         0.6139683         0.1873016
#> X002_BA5_01_28002         0.0000000         0.0000000         0.0000000
#> X003_BA6_01_28003         0.0000000         0.0000000         0.0000000
#> X004_BA7_01_28004         0.0000000         0.0000000         0.0000000
#> X005_BA8_01_28005         0.0000000         0.0000000         0.0000000
#> X011_BB6_01_28012         0.0000000         0.0000000         0.0000000
#> X012_BB7_01_28013         0.0000000         0.0000000         0.0000000
#> X013_BB8_01_28014         0.0000000         0.0000000         0.0000000
#>                   X125_BB3_01_28140 X131_BC1_01_28147 X132_BC2_01_28148
#> X001_BA4_01_28001         0.0000000         0.0000000         0.0000000
#> X002_BA5_01_28002         0.0000000         0.0000000         0.0000000
#> X003_BA6_01_28003         0.0000000         0.0000000         0.0000000
#> X004_BA7_01_28004         0.0000000         0.0000000         0.0000000
#> X005_BA8_01_28005         0.0000000         0.0000000         0.0000000
#> X011_BB6_01_28012         0.0000000         0.0000000         0.0000000
#> X012_BB7_01_28013         0.0000000         0.0000000         0.0000000
#> X013_BB8_01_28014         0.0000000         0.0000000         0.0000000
#>                   X133_BC3_01_28149 X134_BC4_01_28150 X135_BC5_01_28151
#> X001_BA4_01_28001         0.0000000         0.0000000         0.0000000
#> X002_BA5_01_28002         0.0000000         0.0000000         0.0000000
#> X003_BA6_01_28003         0.0000000         0.0000000         0.0000000
#> X004_BA7_01_28004         0.0000000         0.0000000         0.0000000
#> X005_BA8_01_28005         0.0000000         0.0000000         0.0000000
#> X011_BB6_01_28012         0.0000000         0.0000000         0.0000000
#> X012_BB7_01_28013         0.0000000         0.0000000         0.0000000
#> X013_BB8_01_28014         0.0000000         0.0000000         0.0000000
#>                   X136_BC6_01_28152 X137_BC7_01_28153 X138_BC8_01_28154
#> X001_BA4_01_28001         0.0000000         0.0000000         0.0000000
#> X002_BA5_01_28002         0.0000000         0.0000000         0.0000000
#> X003_BA6_01_28003         0.0000000         0.0000000         0.0000000
#> X004_BA7_01_28004         0.0000000         0.0000000         0.0000000
#> X005_BA8_01_28005         0.0000000         0.0000000         0.0000000
#> X011_BB6_01_28012         0.0000000         0.0000000         0.0000000
#> X012_BB7_01_28013         0.0000000         0.0000000         0.0000000
#> X013_BB8_01_28014         0.0000000         0.0000000         0.0000000
#>                   X139_BD1_01_28155 X140_BD2_01_28156 X146_BD8_01_28163
#> X001_BA4_01_28001         0.0000000         0.0000000         0.0000000
#> X002_BA5_01_28002         0.0000000         0.0000000         0.0000000
#> X003_BA6_01_28003         0.0000000         0.0000000         0.0000000
#> X004_BA7_01_28004         0.0000000         0.0000000         0.0000000
#> X005_BA8_01_28005         0.0000000         0.0000000         0.0000000
#> X011_BB6_01_28012         0.0000000         0.0000000         0.0000000
#> X012_BB7_01_28013         0.0000000         0.0000000         0.0000000
#> X013_BB8_01_28014         0.0000000         0.0000000         0.0000000
#>                   X147_BE1_01_28164 X148_BE2_01_28165 X149_BE3_01_28166
#> X001_BA4_01_28001         0.0000000         0.0000000         0.0000000
#> X002_BA5_01_28002         0.0000000         0.0000000         0.0000000
#> X003_BA6_01_28003         0.0000000         0.0000000         0.0000000
#> X004_BA7_01_28004         0.0000000         0.0000000         0.0000000
#> X005_BA8_01_28005         0.0000000         0.5973016         0.0000000
#> X011_BB6_01_28012         0.0000000         0.0000000         0.0000000
#> X012_BB7_01_28013         0.0000000         0.0000000         0.0000000
#> X013_BB8_01_28014         0.0000000         0.0000000         0.0000000
#>                   X150_BE4_01_28167 X151_BE5_01_28169 X152_BE6_01_28170
#> X001_BA4_01_28001         0.0000000         0.0000000         0.0000000
#> X002_BA5_01_28002         0.0000000         0.0000000         0.0000000
#> X003_BA6_01_28003         0.0000000         0.0000000         0.0000000
#> X004_BA7_01_28004         0.0000000         0.4506349         0.0000000
#> X005_BA8_01_28005         0.0000000         0.0000000         0.0000000
#> X011_BB6_01_28012         0.0000000         0.0000000         0.0000000
#> X012_BB7_01_28013         0.0000000         0.0000000         0.0000000
#> X013_BB8_01_28014         0.0000000         0.0000000         0.0000000
#>                   X153_BE7_01_28171 X154_BE8_01_28172 X155_RA1_01_28173
#> X001_BA4_01_28001         0.0000000         0.1963492         0.0000000
#> X002_BA5_01_28002         0.0000000         0.0000000         0.0000000
#> X003_BA6_01_28003         0.0000000         0.0000000         0.0000000
#> X004_BA7_01_28004         0.0000000         0.0000000         0.0000000
#> X005_BA8_01_28005         0.0000000         0.0000000         0.3939683
#> X011_BB6_01_28012         0.0000000         0.0000000         0.0000000
#> X012_BB7_01_28013         0.0000000         0.0000000         0.0000000
#> X013_BB8_01_28014         0.0000000         0.0000000         0.0000000
#>                   X161_RA7_01_28180 X162_RA8_01_28181 X163_RB1_01_28182
#> X001_BA4_01_28001         0.0000000         0.0000000         0.0000000
#> X002_BA5_01_28002         0.0000000         0.0000000         0.0000000
#> X003_BA6_01_28003         0.0000000         0.0000000         0.0000000
#> X004_BA7_01_28004         0.0000000         0.0000000         0.0000000
#> X005_BA8_01_28005         0.0000000         0.0000000         0.0000000
#> X011_BB6_01_28012         0.0000000         0.0000000         0.0000000
#> X012_BB7_01_28013         0.0000000         0.0000000         0.0000000
#> X013_BB8_01_28014         0.0000000         0.0000000         0.0000000
#>                   X164_RB2_01_28183 X165_RB3_01_28184 X166_RB4_01_28185
#> X001_BA4_01_28001         0.0000000         0.0000000         0.0000000
#> X002_BA5_01_28002         0.0000000         0.0000000         0.0000000
#> X003_BA6_01_28003         0.0000000         0.0000000         0.0000000
#> X004_BA7_01_28004         0.0000000         0.0000000         0.0000000
#> X005_BA8_01_28005         0.0000000         0.0000000         0.0000000
#> X011_BB6_01_28012         0.0000000         0.0000000         0.0000000
#> X012_BB7_01_28013         0.0000000         0.0000000         0.0000000
#> X013_BB8_01_28014         0.0000000         0.0000000         0.4923016
#>                   X167_RB5_01_28186 X168_RB6_01_28187 X169_RB7_01_28188
#> X001_BA4_01_28001         0.0000000         0.0000000         0.0000000
#> X002_BA5_01_28002         0.0000000         0.0000000         0.0000000
#> X003_BA6_01_28003         0.0000000         0.0000000         0.0000000
#> X004_BA7_01_28004         0.0000000         0.0000000         0.0000000
#> X005_BA8_01_28005         0.5673016         0.0000000         0.0000000
#> X011_BB6_01_28012         0.0000000         0.0000000         0.0000000
#> X012_BB7_01_28013         0.0000000         0.0000000         0.0000000
#> X013_BB8_01_28014         0.0000000         0.0000000         0.0000000
#>                   X170_RB8_01_28189 X176_RC6_01_28196 X177_RC7_01_28197
#> X001_BA4_01_28001         0.0000000         0.0000000         0.0000000
#> X002_BA5_01_28002         0.0000000         0.0000000         0.0000000
#> X003_BA6_01_28003         0.6839683         0.0000000         0.0000000
#> X004_BA7_01_28004         0.0000000         0.0000000         0.0000000
#> X005_BA8_01_28005         0.0000000         0.0000000         0.0000000
#> X011_BB6_01_28012         0.7039683         0.0000000         0.0000000
#> X012_BB7_01_28013         0.0000000         0.0000000         0.0000000
#> X013_BB8_01_28014         0.0000000         0.0000000         0.0000000
#>                   X178_RC8_01_28198 X179_RD1_01_28199 X180_RD2_01_28200
#> X001_BA4_01_28001         0.0000000         0.0000000         0.0000000
#> X002_BA5_01_28002         0.0000000         0.0000000         0.0000000
#> X003_BA6_01_28003         0.0000000         0.0000000         0.0000000
#> X004_BA7_01_28004         0.0000000         0.0000000         0.0000000
#> X005_BA8_01_28005         0.0000000         0.0000000         0.0000000
#> X011_BB6_01_28012         0.0000000         0.0000000         0.0000000
#> X012_BB7_01_28013         0.0000000         0.0000000         0.0000000
#> X013_BB8_01_28014         0.0000000         0.0000000         0.0000000
#>  [ reached getOption("max.print") -- omitted 112 rows ]