Create an adjacency matrix based on score-based structure learning algorithm

`bayes` infers an adjacency matrix using score-based structure learning algorithm `boot.strength` from the `bnlearn` package. `bayes` extracts then the reported connections from running the `boot.strength` function and assigns the strengths of the arcs of the Bayesian connections to an adjacency matrix. `bayes` returns this weighted adjacency matrix.

bayes(x, algorithm = "tabu", R = 100, ...)

Arguments

x

`matrix` where columns are the samples and the rows are features (metabolites), cell entries are intensity values

algorithm

`character`, structure learning to be applied to the bootstrap replicates (default is `"tabu"`)

R

`numeric`, number of bootstrap replicates

...

parameters passed to `boot.strength`

Value

`matrix` with edges inferred from score-based structure learning algorithm `boot.strength`

Details

`boot.strength` measures the strength of the probabilistic relationships by the arcs of a Bayesian network, as learned from bootstrapped data. By default `bayes` uses the Tabu greedy search.

For use of the parameters used in the `boot.strength` function, refer to `?bnlearn::boot.strength`. For further information see also Friedman et al. (1999) and Scutari and Nagarajan (2001).

References

Friedman et al. (1999): Data Analysis with Bayesian Networks: A Bootstrap Approach. Proceedings of the 15th Annual Conference on Uncertainty in Artificial Intelligence, 196-201.

Scutari and Nagarajan (2011): On Identifying Significant Edges in Graphical Models. Proceedings of the Workshop Probabilistic Problem Solving in Biomedicine of the 13th Artificial Intelligence in Medicine Conference, 15-27.

Author

Thomas Naake, thomasnaake@googlemail.com

Examples

data("x_test", package = "MetNet")
x <- x_test[1:10, 3:ncol(x_test)]
x <- as.matrix(x)
bayes(x, algorithm = "tabu", R = 100)
#>        x9485 x7449 x11179 x11374 x13007 x14826 x16257 x10804 x12669 x14553
#> x9485   0.00  1.00   0.69   1.00   0.81   0.34   0.17   0.21   0.36   0.36
#> x7449   1.00  0.00   1.00   0.79   0.70   0.29   0.43   0.82   0.19   0.23
#> x11179  0.69  1.00   0.00   0.78   1.00   0.96   0.21   0.41   0.15   0.32
#> x11374  1.00  0.79   0.78   0.00   0.50   0.62   0.54   0.20   0.16   0.37
#> x13007  0.81  0.70   1.00   0.50   0.00   1.00   0.95   0.15   0.25   0.45
#> x14826  0.34  0.29   0.96   0.62   1.00   0.00   1.00   0.13   0.30   0.59
#> x16257  0.17  0.43   0.21   0.54   0.95   1.00   0.00   0.18   0.54   0.89
#> x10804  0.21  0.82   0.41   0.20   0.15   0.13   0.18   0.00   1.00   1.00
#> x12669  0.36  0.19   0.15   0.16   0.25   0.30   0.54   1.00   0.00   1.00
#> x14553  0.36  0.23   0.32   0.37   0.45   0.59   0.89   1.00   1.00   0.00