Create an adjacency matrix based on correlation

`correlation` infers an adjacency matrix using correlation using the `corr.test` function (from the `psych` package) or `partialCorrelation`. `correlation` extracts the reported pair-wise correlation coefficients from the function `corr.test` and `partialCorrelation` and will return the weighted adjacency matrix of the correlation coefficients, together with the associated p-values.

correlation(x, method = "pearson", p.adjust = "none", ...)

Arguments

x

`matrix`, where columns are the samples and the rows are features (metabolites), cell entries are intensity values

method

`character`, either "pearson", "spearman", "pearson_partial", "spearman_partial", or "ggm".

p.adjust

`character`, method of p-value adjustment passed to `p.adjust`

...

additional arguments passed to `corr.test` or `partialCorrelation`

Value

`list` containing two matrices, the first matrix contains correlation coefficients and the second matrix contains the corresponding p-values as obtained from the correlation algorithms `corr.test` or `partialCorrelation` (depending on the chosen `method`) and optionally the adjusted p.values (argument `p.adjust`)

Details

If `"pearson"` or `"spearman"` is used as a `method`, the function `corr.test` from `psych` will be employed.

If `"ggm"` is used as a `method`, the function `ggm.estimate.pcor` from `GeneNet` will be employed.

If `"pearson_partial"` or `"spearman_partial"` is used as a `method` the function `partialCorrelation` will be employed.

`method` will be passed to argument `method` in `corr.test` (in the case of `"pearson"` or `"spearman"`) or to `method` in `partialCorrelation` (`"pearson"` and `"spearman"` for `"pearson_partial"` and `"spearman_partial"`, respectively).

Author

Thomas Naake, thomasnaake@googlemail.com, Liesa Salzer, liesa.salzer@helmholtz-muenchen.de

Examples

data("x_test", package = "MetNet")
x <- x_test[1:10, 3:ncol(x_test)]
x <- as.matrix(x)
correlation(x, method = "pearson")
#> $r
#>            x9485     x7449     x11179    x11374    x13007    x14826      x16257
#> x9485  1.0000000 0.8000351 0.78391122 0.9403697 0.8510633 0.6986100  0.27631528
#> x7449  0.8000351 1.0000000 0.92406116 0.6044475 0.7470669 0.4622945  0.01600220
#> x11179 0.7839112 0.9240612 1.00000000 0.5948374 0.8648860 0.5646757  0.01969613
#> x11374 0.9403697 0.6044475 0.59483740 1.0000000 0.7782440 0.7367013  0.35634911
#> x13007 0.8510633 0.7470669 0.86488604 0.7782440 1.0000000 0.8600956  0.32016094
#> x14826 0.6986100 0.4622945 0.56467568 0.7367013 0.8600956 1.0000000  0.68398191
#> x16257 0.2763153 0.0160022 0.01969613 0.3563491 0.3201609 0.6839819  1.00000000
#> x10804 0.5578105 0.7620814 0.74124973 0.4150268 0.6205390 0.3859232 -0.09495813
#> x12669 0.5059605 0.6519108 0.66648908 0.4345848 0.6217924 0.4400734 -0.12757798
#> x14553 0.3993292 0.5008692 0.54831042 0.3736941 0.5383033 0.3926268 -0.19360513
#>             x10804     x12669     x14553
#> x9485   0.55781050  0.5059605  0.3993292
#> x7449   0.76208143  0.6519108  0.5008692
#> x11179  0.74124973  0.6664891  0.5483104
#> x11374  0.41502679  0.4345848  0.3736941
#> x13007  0.62053902  0.6217924  0.5383033
#> x14826  0.38592321  0.4400734  0.3926268
#> x16257 -0.09495813 -0.1275780 -0.1936051
#> x10804  1.00000000  0.8902988  0.7161565
#> x12669  0.89029884  1.0000000  0.9398149
#> x14553  0.71615645  0.9398149  1.0000000
#> 
#> $p
#>               x9485        x7449       x11179       x11374       x13007
#> x9485  0.000000e+00 5.995208e-28 3.490539e-26 4.256342e-57 8.292921e-35
#> x7449  5.995208e-28 0.000000e+00 4.121436e-51 2.672329e-13 1.154774e-22
#> x11179 3.490539e-26 4.121436e-51 0.000000e+00 7.846964e-13 4.043036e-37
#> x11374 4.256342e-57 2.672329e-13 7.846964e-13 0.000000e+00 1.342061e-25
#> x13007 8.292921e-35 1.154774e-22 4.043036e-37 1.342061e-25 0.000000e+00
#> x14826 7.265344e-19 1.063908e-07 1.848667e-11 8.811483e-22 2.728331e-36
#> x16257 2.251303e-03 8.622808e-01 8.309193e-01 6.477481e-05 3.641525e-04
#> x10804 3.632235e-11 5.073495e-24 3.655654e-22 2.438916e-06 4.050218e-14
#> x12669 3.749261e-09 7.363341e-16 9.698559e-17 7.065826e-07 3.480764e-14
#> x14553 6.240493e-06 5.675659e-09 9.019640e-11 2.621014e-05 2.280927e-10
#>              x14826       x16257       x10804       x12669       x14553
#> x9485  7.265344e-19 2.251303e-03 3.632235e-11 3.749261e-09 6.240493e-06
#> x7449  1.063908e-07 8.622808e-01 5.073495e-24 7.363341e-16 5.675659e-09
#> x11179 1.848667e-11 8.309193e-01 3.655654e-22 9.698559e-17 9.019640e-11
#> x11374 8.811483e-22 6.477481e-05 2.438916e-06 7.065826e-07 2.621014e-05
#> x13007 2.728331e-36 3.641525e-04 4.050218e-14 3.480764e-14 2.280927e-10
#> x14826 0.000000e+00 7.293261e-18 1.341654e-05 4.920649e-07 9.188041e-06
#> x16257 7.293261e-18 0.000000e+00 3.022316e-01 1.649544e-01 3.411338e-02
#> x10804 1.341654e-05 3.022316e-01 0.000000e+00 4.002774e-42 3.777603e-20
#> x12669 4.920649e-07 1.649544e-01 4.002774e-42 0.000000e+00 7.232093e-57
#> x14553 9.188041e-06 3.411338e-02 3.777603e-20 7.232093e-57 0.000000e+00
#>