Calculate the partial correlation and p-values

`partialCorrelation` infers an adjacency matrix of partial correlation values and associated p-values using using the `cor2pcor` function (from the `corpcor` package). `partialCorrelation` calculates the p-values from the number of samples (`n`) and the number of controlling variables (`g`). The function will return a list containing the weighted adjacency matrix of the correlation values, together with the associated p-values.

partialCorrelation(x, method = "pearson", ...)

Arguments

x

`matrix`, where columns are the features (metabolites) and the rows are samples, cell entries are intensity values

method

`character`, either "pearson", "spearman"

...

further arguments passed to `cor` from `base` or `cor2pcor` from `corpcor`

Value

`list` containing two matrices, the first matrix contains correlation coefficients and the second matrix contains the corresponding p-values

Details

The correlation coefficients \(r_{ij|S}\) are obtained from `cor2pcor` (`corpcor` package).

The t-values are calculated via

\(t_{ij|S} = r_{ij|S} \cdot \sqrt{\frac{n-2-g}{1-r_{ij|S}^2}} \), where $n$ are the number of samples and $g$ the number of controlling variables (number of features - 2).

The p-values are calculated as follows \(p_{ij|S} = 2 \cdot pt(-abs(t_{ij|S}), df = n - 2 - g)\)

Author

Thomas Naake, thomasnaake@googlemail.com

Examples

data("x_test", package = "MetNet")
x <- x_test[, 3:ncol(x_test)]
x <- as.matrix(x)
x <- t(x)
partialCorrelation(x, use = "pairwise", method = "pearson")
#> Estimating optimal shrinkage intensity lambda (correlation matrix): 0.1109 
#> 
#> $r
#>               [,1]          [,2]          [,3]          [,4]          [,5]
#>  [1,]  1.000000000  0.1698629980  0.1130361820  0.4356753982  0.1282482693
#>  [2,]  0.169862998  1.0000000000  0.2082066776  0.0550195247  0.0950850145
#>  [3,]  0.113036182  0.2082066776  1.0000000000 -0.0053435474  0.1420908524
#>  [4,]  0.435675398  0.0550195247 -0.0053435474  1.0000000000  0.1307649039
#>  [5,]  0.128248269  0.0950850145  0.1420908524  0.1307649039  1.0000000000
#>  [6,]  0.073888576 -0.0651724049  0.0343911731  0.1994439499  0.2919034963
#>  [7,] -0.004145819 -0.0389352324 -0.0282639229  0.0286533947  0.0674720059
#>  [8,]  0.029907941  0.1040126023  0.0847542984 -0.0378700587  0.0336019782
#>  [9,]  0.006730555  0.0294313643  0.0246307184  0.0089796342  0.0253358027
#> [10,] -0.010150717 -0.0161631773 -0.0098811411  0.0272884162  0.0160745018
#> [11,]  0.026804715  0.1322011890  0.1473184608 -0.0941895329  0.0859580651
#> [12,]  0.005696758  0.0450601094  0.0750005432 -0.0289839248  0.0835310154
#> [13,] -0.005451755 -0.0102084095  0.0286827155  0.0128123531  0.0756768341
#> [14,]  0.087537122  0.2144261298  0.2034007356 -0.0608898528  0.0997049555
#> [15,]  0.036291742  0.0869813900  0.0926200994 -0.0068150727  0.0735652494
#> [16,] -0.002867884  0.0078528026  0.0271642409  0.0098109872  0.0545812973
#> [17,] -0.030022911 -0.0189692794 -0.0126046254  0.0004718421  0.0116465314
#> [18,]  0.157609229  0.0571457732 -0.0102267400  0.1895374234 -0.0368910632
#> [19,]  0.016897515  0.0242902807 -0.0033737969  0.0304859402 -0.0188184420
#> [20,]  0.027538281  0.0132497507 -0.0069334094  0.0275927462  0.0010106258
#> [21,]  0.025492016  0.0196923419 -0.0044490139  0.0211363811  0.0012158308
#> [22,]  0.022316346 -0.0099085364 -0.0317862355  0.0793141210 -0.0167453228
#> [23,]  0.009234133  0.0102567299  0.0090461463 -0.0225322938  0.0467177135
#> [24,] -0.036071432 -0.0357417580 -0.0220783039 -0.0295283094  0.0135629146
#> [25,] -0.044898986  0.0005629342  0.0192411525 -0.0988061476  0.0388024404
#> [26,]  0.003496715 -0.0310875951 -0.0499231654  0.0699337072 -0.0255076877
#> [27,] -0.024887174 -0.0141667410 -0.0250865049  0.0160388527 -0.0137962633
#> [28,]  0.060227100 -0.0087199141 -0.0073321626  0.0630011438  0.0241068603
#> [29,] -0.075110851 -0.1161379163 -0.0525061675  0.0148787322  0.0535813402
#> [30,] -0.045716225 -0.0396916830 -0.0466631048  0.0087788223 -0.0256159257
#> [31,] -0.021290689 -0.0072903630 -0.0241085553  0.0086411469 -0.0217228632
#> [32,]  0.053188514 -0.0198748700  0.0133401694  0.1110327096  0.0531292123
#> [33,]  0.078307913  0.0099809698  0.0381764410  0.1355517283  0.0746465754
#> [34,]  0.031890221  0.0090548495 -0.0128724191  0.0409525036  0.0006854315
#> [35,]  0.017240469 -0.0214271654  0.0005732978  0.0514856415  0.0457015417
#> [36,]  0.003449230 -0.0053472524  0.0269261695  0.0102085283  0.0577238783
#>               [,6]         [,7]         [,8]         [,9]         [,10]
#>  [1,]  0.073888576 -0.004145819  0.029907941  0.006730555 -0.0101507165
#>  [2,] -0.065172405 -0.038935232  0.104012602  0.029431364 -0.0161631773
#>  [3,]  0.034391173 -0.028263923  0.084754298  0.024630718 -0.0098811411
#>  [4,]  0.199443950  0.028653395 -0.037870059  0.008979634  0.0272884162
#>  [5,]  0.291903496  0.067472006  0.033601978  0.025335803  0.0160745018
#>  [6,]  1.000000000  0.405003356 -0.055381517  0.025298098  0.0603466270
#>  [7,]  0.405003356  1.000000000 -0.066543984 -0.032259865 -0.0229113979
#>  [8,] -0.055381517 -0.066543984  1.000000000  0.078169161  0.0431900263
#>  [9,]  0.025298098 -0.032259865  0.078169161  1.000000000  0.1071869695
#> [10,]  0.060346627 -0.022911398  0.043190026  0.107186969  1.0000000000
#> [11,] -0.036755699 -0.091353441  0.103603710  0.020021023 -0.0211893372
#> [12,]  0.067734546 -0.053550517  0.051011045  0.054704554  0.0504303474
#> [13,]  0.113367649 -0.043536629  0.024037562  0.071080803  0.0903655560
#> [14,] -0.045147089 -0.063733481  0.117148494  0.034338740 -0.0107998411
#> [15,]  0.042781287 -0.025681777  0.061133601  0.077016158  0.0812844434
#> [16,]  0.103418567 -0.001302880  0.028170029  0.097250694  0.1334832565
#> [17,]  0.069437764  0.005274415  0.012439369  0.104946488  0.1526555805
#> [18,] -0.087479768 -0.017650382  0.078716058  0.015294035 -0.0132349320
#> [19,] -0.079899593 -0.057040321  0.086145326  0.056703141  0.0410321500
#> [20,]  0.060044995  0.080374768  0.037225129 -0.010797741 -0.0421171600
#> [21,]  0.061632294  0.087584659  0.041034892 -0.008441139 -0.0422916663
#> [22,] -0.026520551 -0.022106673  0.054135003  0.042287964  0.0313104360
#> [23,]  0.101824643  0.106673641 -0.078055815 -0.059579963 -0.0787795051
#> [24,]  0.091576422  0.070401886 -0.024382203 -0.020690284 -0.0323359086
#> [25,]  0.119080392  0.084636029  0.026501406 -0.019361446 -0.0503680315
#> [26,]  0.009529414  0.011969205  0.038286400  0.044365132  0.0421172008
#> [27,] -0.032643592 -0.035682677  0.055000774  0.059351626  0.0570974445
#> [28,]  0.097580666  0.097815088  0.006346564 -0.043134838 -0.0653793863
#> [29,]  0.139872149 -0.137702139 -0.021496642  0.013167322  0.0490091181
#> [30,] -0.032328761 -0.030515266  0.011358200  0.052005577  0.0655649372
#> [31,] -0.055109056 -0.046380363  0.050794367  0.057299102  0.0560693438
#> [32,] -0.007722086 -0.098997102 -0.014787458 -0.004396989  0.0121891367
#> [33,]  0.010562501 -0.074513084 -0.028637216 -0.014711704 -0.0004886037
#> [34,]  0.069286849 -0.016467401  0.028976199  0.026379582  0.0733532504
#> [35,]  0.073333386  0.016017417 -0.033455708 -0.054465803 -0.0697485276
#> [36,]  0.056113328  0.003159288 -0.039193564 -0.057363577 -0.0725313472
#>              [,11]         [,12]        [,13]         [,14]         [,15]
#>  [1,]  0.026804715  0.0056967578 -0.005451755  0.0875371223  0.0362917420
#>  [2,]  0.132201189  0.0450601094 -0.010208409  0.2144261298  0.0869813900
#>  [3,]  0.147318461  0.0750005432  0.028682715  0.2034007356  0.0926200994
#>  [4,] -0.094189533 -0.0289839248  0.012812353 -0.0608898528 -0.0068150727
#>  [5,]  0.085958065  0.0835310154  0.075676834  0.0997049555  0.0735652494
#>  [6,] -0.036755699  0.0677345458  0.113367649 -0.0451470893  0.0427812875
#>  [7,] -0.091353441 -0.0535505174 -0.043536629 -0.0637334813 -0.0256817768
#>  [8,]  0.103603710  0.0510110451  0.024037562  0.1171484939  0.0611336010
#>  [9,]  0.020021023  0.0547045541  0.071080803  0.0343387404  0.0770161576
#> [10,] -0.021189337  0.0504303474  0.090365556 -0.0107998411  0.0812844434
#> [11,]  1.000000000  0.1328733306  0.068612547  0.1761546990  0.0504472785
#> [12,]  0.132873331  1.0000000000  0.137405175  0.0726706616  0.0683108822
#> [13,]  0.068612547  0.1374051748  1.000000000  0.0140445523  0.0690895338
#> [14,]  0.176154699  0.0726706616  0.014044552  1.0000000000  0.0964622221
#> [15,]  0.050447278  0.0683108822  0.069089534  0.0964622221  1.0000000000
#> [16,] -0.011206692  0.0642885145  0.098533106  0.0182170930  0.1047933203
#> [17,] -0.048228031  0.0495504174  0.093548331 -0.0209181185  0.1022816040
#> [18,]  0.025797505 -0.0318162540 -0.052500091  0.0163593494 -0.0314080738
#> [19,]  0.027467246 -0.0094633240 -0.019386355  0.0196675233  0.0198419002
#> [20,]  0.033009455 -0.0179494905 -0.045282690  0.0069145299 -0.0322823881
#> [21,]  0.032241138 -0.0170083857 -0.045277178  0.0058006468 -0.0316712740
#> [22,]  0.016878508  0.0001059202 -0.004418737 -0.0172060026  0.0025580404
#> [23,] -0.089048198 -0.0210663821 -0.030439912 -0.0342456765  0.0187087488
#> [24,] -0.002812545 -0.0029168608 -0.013224004 -0.0083379075 -0.0056504502
#> [25,]  0.066336169  0.0104386605 -0.021804879  0.0373709103 -0.0168098490
#> [26,]  0.000314226  0.0009941431  0.003579169 -0.0353246232  0.0018757859
#> [27,]  0.016374164  0.0116092816  0.011569417 -0.0147942163  0.0258218865
#> [28,]  0.054204158 -0.0200979593 -0.049165098  0.0134280130 -0.0543054573
#> [29,]  0.172824221  0.2311777951  0.265131911 -0.0925675335 -0.0619928248
#> [30,] -0.042666015 -0.0117972847  0.004019595 -0.0339712003  0.0401172879
#> [31,] -0.006552320 -0.0058366036 -0.001933514 -0.0066568098  0.0317633930
#> [32,] -0.051065663 -0.0280974041 -0.006061213 -0.0124340193  0.0158992700
#> [33,] -0.051553856 -0.0287392680 -0.012050911  0.0026647296  0.0250159205
#> [34,] -0.112985606 -0.0970447628 -0.060122138 -0.0459101995  0.0001395994
#> [35,] -0.003226755 -0.0178253908 -0.023196770 -0.0270593837 -0.0404793989
#> [36,] -0.004088750 -0.0132192909 -0.019891796 -0.0006458635 -0.0264276011
#>              [,16]         [,17]        [,18]         [,19]         [,20]
#>  [1,] -0.002867884 -0.0300229111  0.157609229  0.0168975148  0.0275382811
#>  [2,]  0.007852803 -0.0189692794  0.057145773  0.0242902807  0.0132497507
#>  [3,]  0.027164241 -0.0126046254 -0.010226740 -0.0033737969 -0.0069334094
#>  [4,]  0.009810987  0.0004718421  0.189537423  0.0304859402  0.0275927462
#>  [5,]  0.054581297  0.0116465314 -0.036891063 -0.0188184420  0.0010106258
#>  [6,]  0.103418567  0.0694377636 -0.087479768 -0.0798995925  0.0600449948
#>  [7,] -0.001302880  0.0052744154 -0.017650382 -0.0570403210  0.0803747679
#>  [8,]  0.028170029  0.0124393690  0.078716058  0.0861453264  0.0372251290
#>  [9,]  0.097250694  0.1049464877  0.015294035  0.0567031414 -0.0107977409
#> [10,]  0.133483257  0.1526555805 -0.013234932  0.0410321500 -0.0421171600
#> [11,] -0.011206692 -0.0482280314  0.025797505  0.0274672457  0.0330094554
#> [12,]  0.064288514  0.0495504174 -0.031816254 -0.0094633240 -0.0179494905
#> [13,]  0.098533106  0.0935483305 -0.052500091 -0.0193863548 -0.0452826897
#> [14,]  0.018217093 -0.0209181185  0.016359349  0.0196675233  0.0069145299
#> [15,]  0.104793320  0.1022816040 -0.031408074  0.0198419002 -0.0322823881
#> [16,]  1.000000000  0.1625663755 -0.061175318  0.0134535619 -0.0559069074
#> [17,]  0.162566376  1.0000000000 -0.063214672  0.0261309664 -0.0585511741
#> [18,] -0.061175318 -0.0632146721  1.000000000  0.0966437975  0.1368081552
#> [19,]  0.013453562  0.0261309664  0.096643798  1.0000000000  0.0586464775
#> [20,] -0.055906907 -0.0585511741  0.136808155  0.0586464775  1.0000000000
#> [21,] -0.056867965 -0.0585848824  0.133742544  0.0587176262  0.1186565770
#> [22,]  0.006001351  0.0235848368  0.080714244  0.1115746125  0.0652182997
#> [23,] -0.003461113  0.0358789443  0.073227332 -0.0009147784  0.0780634816
#> [24,] -0.009777361  0.0049387089 -0.025527811  0.0344523045  0.0616487614
#> [25,] -0.041298005 -0.0536085936  0.024408570  0.0421707099  0.0968911284
#> [26,]  0.018017607  0.0416444629  0.058743762  0.0939233597  0.0603428391
#> [27,]  0.040108289  0.0644335288 -0.014499161  0.1112992598  0.0290088553
#> [28,] -0.080977843 -0.0970666169  0.181251440  0.0170012236  0.1106614346
#> [29,] -0.013439159 -0.0122202781 -0.010626510 -0.0338947362 -0.0321645320
#> [30,]  0.067553563  0.1098825787 -0.078481955  0.1036108089 -0.0005533343
#> [31,]  0.043366472  0.0701946307 -0.001859765  0.1154842487  0.0246701631
#> [32,]  0.024321758  0.0030359567 -0.053257631 -0.0016602617 -0.0705859437
#> [33,]  0.025334725  0.0071755866 -0.086229740 -0.0138714868 -0.0765725133
#> [34,]  0.030763177  0.0373420370  0.151787754  0.0389687105 -0.0238059467
#> [35,] -0.055702241 -0.0882944917 -0.064242550 -0.0420921656  0.0047888329
#> [36,] -0.046085662 -0.0776485176 -0.130703485 -0.0609813799 -0.0320530614
#>              [,21]         [,22]         [,23]        [,24]         [,25]
#>  [1,]  0.025492016  0.0223163462  0.0092341332 -0.036071432 -0.0448989857
#>  [2,]  0.019692342 -0.0099085364  0.0102567299 -0.035741758  0.0005629342
#>  [3,] -0.004449014 -0.0317862355  0.0090461463 -0.022078304  0.0192411525
#>  [4,]  0.021136381  0.0793141210 -0.0225322938 -0.029528309 -0.0988061476
#>  [5,]  0.001215831 -0.0167453228  0.0467177135  0.013562915  0.0388024404
#>  [6,]  0.061632294 -0.0265205512  0.1018246428  0.091576422  0.1190803916
#>  [7,]  0.087584659 -0.0221066725  0.1066736411  0.070401886  0.0846360293
#>  [8,]  0.041034892  0.0541350026 -0.0780558150 -0.024382203  0.0265014057
#>  [9,] -0.008441139  0.0422879644 -0.0595799626 -0.020690284 -0.0193614457
#> [10,] -0.042291666  0.0313104360 -0.0787795051 -0.032335909 -0.0503680315
#> [11,]  0.032241138  0.0168785079 -0.0890481977 -0.002812545  0.0663361690
#> [12,] -0.017008386  0.0001059202 -0.0210663821 -0.002916861  0.0104386605
#> [13,] -0.045277178 -0.0044187370 -0.0304399124 -0.013224004 -0.0218048792
#> [14,]  0.005800647 -0.0172060026 -0.0342456765 -0.008337908  0.0373709103
#> [15,] -0.031671274  0.0025580404  0.0187087488 -0.005650450 -0.0168098490
#> [16,] -0.056867965  0.0060013514 -0.0034611128 -0.009777361 -0.0412980051
#> [17,] -0.058584882  0.0235848368  0.0358789443  0.004938709 -0.0536085936
#> [18,]  0.133742544  0.0807142443  0.0732273323 -0.025527811  0.0244085702
#> [19,]  0.058717626  0.1115746125 -0.0009147784  0.034452305  0.0421707099
#> [20,]  0.118656577  0.0652182997  0.0780634816  0.061648761  0.0968911284
#> [21,]  1.000000000  0.0656605054  0.0736025747  0.054345610  0.0947872154
#> [22,]  0.065660505  1.0000000000  0.0099981749  0.054973600  0.0471690802
#> [23,]  0.073602575  0.0099981749  1.0000000000  0.280067347  0.1509451995
#> [24,]  0.054345610  0.0549735996  0.2800673467  1.000000000  0.1360072422
#> [25,]  0.094787215  0.0471690802  0.1509451995  0.136007242  1.0000000000
#> [26,]  0.061617656  0.1069789482  0.0019290177  0.058391860  0.0448717939
#> [27,]  0.031095798  0.1117640901 -0.0103021397  0.056569699  0.0364483993
#> [28,]  0.109665642  0.0294694058 -0.0195045981  0.032902425  0.0908950037
#> [29,] -0.039380082  0.0081743506 -0.0104651943 -0.009782643  0.0144021115
#> [30,] -0.002994793  0.1067385149  0.1340595603  0.123941686  0.0426338399
#> [31,]  0.024104283  0.1066365081  0.0631473606  0.079813645  0.0440547810
#> [32,] -0.074771539 -0.0227438955  0.0327347373 -0.048160457 -0.0665044780
#> [33,] -0.079460073 -0.0299074551  0.0804904928 -0.031612863 -0.0655040503
#> [34,] -0.027873442 -0.0372621773  0.1714421684 -0.112940994 -0.0845325775
#> [35,]  0.012674411 -0.0253566593 -0.0293468554 -0.041849194 -0.0010745962
#> [36,] -0.029623902 -0.0517939936  0.0058168253 -0.016263893 -0.0065051810
#>               [,26]         [,27]        [,28]         [,29]         [,30]
#>  [1,]  0.0034967148 -0.0248871739  0.060227100 -0.0751108509 -0.0457162248
#>  [2,] -0.0310875951 -0.0141667410 -0.008719914 -0.1161379163 -0.0396916830
#>  [3,] -0.0499231654 -0.0250865049 -0.007332163 -0.0525061675 -0.0466631048
#>  [4,]  0.0699337072  0.0160388527  0.063001144  0.0148787322  0.0087788223
#>  [5,] -0.0255076877 -0.0137962633  0.024106860  0.0535813402 -0.0256159257
#>  [6,]  0.0095294137 -0.0326435920  0.097580666  0.1398721493 -0.0323287606
#>  [7,]  0.0119692046 -0.0356826774  0.097815088 -0.1377021386 -0.0305152661
#>  [8,]  0.0382864005  0.0550007743  0.006346564 -0.0214966420  0.0113582000
#>  [9,]  0.0443651316  0.0593516261 -0.043134838  0.0131673219  0.0520055767
#> [10,]  0.0421172008  0.0570974445 -0.065379386  0.0490091181  0.0655649372
#> [11,]  0.0003142260  0.0163741640  0.054204158  0.1728242206 -0.0426660147
#> [12,]  0.0009941431  0.0116092816 -0.020097959  0.2311777951 -0.0117972847
#> [13,]  0.0035791685  0.0115694169 -0.049165098  0.2651319109  0.0040195951
#> [14,] -0.0353246232 -0.0147942163  0.013428013 -0.0925675335 -0.0339712003
#> [15,]  0.0018757859  0.0258218865 -0.054305457 -0.0619928248  0.0401172879
#> [16,]  0.0180176066  0.0401082889 -0.080977843 -0.0134391585  0.0675535626
#> [17,]  0.0416444629  0.0644335288 -0.097066617 -0.0122202781  0.1098825787
#> [18,]  0.0587437618 -0.0144991608  0.181251440 -0.0106265103 -0.0784819548
#> [19,]  0.0939233597  0.1112992598  0.017001224 -0.0338947362  0.1036108089
#> [20,]  0.0603428391  0.0290088553  0.110661435 -0.0321645320 -0.0005533343
#> [21,]  0.0616176565  0.0310957984  0.109665642 -0.0393800823 -0.0029947931
#> [22,]  0.1069789482  0.1117640901  0.029469406  0.0081743506  0.1067385149
#> [23,]  0.0019290177 -0.0103021397 -0.019504598 -0.0104651943  0.1340595603
#> [24,]  0.0583918600  0.0565696990  0.032902425 -0.0097826431  0.1239416859
#> [25,]  0.0448717939  0.0364483993  0.090895004  0.0144021115  0.0426338399
#> [26,]  1.0000000000  0.1080132331  0.030344336  0.0055799416  0.1069383991
#> [27,]  0.1080132331  1.0000000000 -0.029546565 -0.0002808827  0.1512623862
#> [28,]  0.0303443356 -0.0295465645  1.000000000 -0.0108529447 -0.0684998263
#> [29,]  0.0055799416 -0.0002808827 -0.010852945  1.0000000000 -0.0591328162
#> [30,]  0.1069383991  0.1512623862 -0.068499826 -0.0591328162  1.0000000000
#> [31,]  0.0987909516  0.1316911601 -0.033611282 -0.0348005138  0.1677081226
#> [32,] -0.0438839009 -0.0057361651 -0.085639046 -0.0887969044  0.0146068004
#> [33,] -0.0492371472 -0.0063131973 -0.092483844 -0.1153680024  0.0277696929
#> [34,] -0.0682464949 -0.0555979616 -0.074269813  0.2862871379 -0.0575648238
#> [35,] -0.0318729227 -0.0329026741 -0.013394707 -0.0842083911 -0.0537757897
#> [36,] -0.0586095248 -0.0380989698 -0.052904939 -0.0484719121 -0.0378776453
#>              [,31]        [,32]         [,33]         [,34]         [,35]
#>  [1,] -0.021290689  0.053188514  0.0783079135  0.0318902207  0.0172404687
#>  [2,] -0.007290363 -0.019874870  0.0099809698  0.0090548495 -0.0214271654
#>  [3,] -0.024108555  0.013340169  0.0381764410 -0.0128724191  0.0005732978
#>  [4,]  0.008641147  0.111032710  0.1355517283  0.0409525036  0.0514856415
#>  [5,] -0.021722863  0.053129212  0.0746465754  0.0006854315  0.0457015417
#>  [6,] -0.055109056 -0.007722086  0.0105625015  0.0692868492  0.0733333859
#>  [7,] -0.046380363 -0.098997102 -0.0745130843 -0.0164674010  0.0160174166
#>  [8,]  0.050794367 -0.014787458 -0.0286372159  0.0289761986 -0.0334557082
#>  [9,]  0.057299102 -0.004396989 -0.0147117044  0.0263795817 -0.0544658032
#> [10,]  0.056069344  0.012189137 -0.0004886037  0.0733532504 -0.0697485276
#> [11,] -0.006552320 -0.051065663 -0.0515538557 -0.1129856063 -0.0032267546
#> [12,] -0.005836604 -0.028097404 -0.0287392680 -0.0970447628 -0.0178253908
#> [13,] -0.001933514 -0.006061213 -0.0120509111 -0.0601221378 -0.0231967700
#> [14,] -0.006656810 -0.012434019  0.0026647296 -0.0459101995 -0.0270593837
#> [15,]  0.031763393  0.015899270  0.0250159205  0.0001395994 -0.0404793989
#> [16,]  0.043366472  0.024321758  0.0253347248  0.0307631775 -0.0557022406
#> [17,]  0.070194631  0.003035957  0.0071755866  0.0373420370 -0.0882944917
#> [18,] -0.001859765 -0.053257631 -0.0862297405  0.1517877542 -0.0642425496
#> [19,]  0.115484249 -0.001660262 -0.0138714868  0.0389687105 -0.0420921656
#> [20,]  0.024670163 -0.070585944 -0.0765725133 -0.0238059467  0.0047888329
#> [21,]  0.024104283 -0.074771539 -0.0794600734 -0.0278734417  0.0126744106
#> [22,]  0.106636508 -0.022743895 -0.0299074551 -0.0372621773 -0.0253566593
#> [23,]  0.063147361  0.032734737  0.0804904928  0.1714421684 -0.0293468554
#> [24,]  0.079813645 -0.048160457 -0.0316128632 -0.1129409944 -0.0418491938
#> [25,]  0.044054781 -0.066504478 -0.0655040503 -0.0845325775 -0.0010745962
#> [26,]  0.098790952 -0.043883901 -0.0492371472 -0.0682464949 -0.0318729227
#> [27,]  0.131691160 -0.005736165 -0.0063131973 -0.0555979616 -0.0329026741
#> [28,] -0.033611282 -0.085639046 -0.0924838439 -0.0742698129 -0.0133947073
#> [29,] -0.034800514 -0.088796904 -0.1153680024  0.2862871379 -0.0842083911
#> [30,]  0.167708123  0.014606800  0.0277696929 -0.0575648238 -0.0537757897
#> [31,]  1.000000000  0.006011546  0.0056748527 -0.0083386156 -0.0482977776
#> [32,]  0.006011546  1.000000000  0.2112831288  0.2166207075  0.0899480643
#> [33,]  0.005674853  0.211283129  1.0000000000  0.1621358420  0.0846857737
#> [34,] -0.008338616  0.216620707  0.1621358420  1.0000000000 -0.0208975170
#> [35,] -0.048297778  0.089948064  0.0846857737 -0.0208975170  1.0000000000
#> [36,] -0.047789892  0.090352176  0.1002936177 -0.0227585045  0.1980609811
#>               [,36]
#>  [1,]  0.0034492301
#>  [2,] -0.0053472524
#>  [3,]  0.0269261695
#>  [4,]  0.0102085283
#>  [5,]  0.0577238783
#>  [6,]  0.0561133278
#>  [7,]  0.0031592877
#>  [8,] -0.0391935643
#>  [9,] -0.0573635772
#> [10,] -0.0725313472
#> [11,] -0.0040887505
#> [12,] -0.0132192909
#> [13,] -0.0198917956
#> [14,] -0.0006458635
#> [15,] -0.0264276011
#> [16,] -0.0460856619
#> [17,] -0.0776485176
#> [18,] -0.1307034854
#> [19,] -0.0609813799
#> [20,] -0.0320530614
#> [21,] -0.0296239023
#> [22,] -0.0517939936
#> [23,]  0.0058168253
#> [24,] -0.0162638927
#> [25,] -0.0065051810
#> [26,] -0.0586095248
#> [27,] -0.0380989698
#> [28,] -0.0529049395
#> [29,] -0.0484719121
#> [30,] -0.0378776453
#> [31,] -0.0477898924
#> [32,]  0.0903521760
#> [33,]  0.1002936177
#> [34,] -0.0227585045
#> [35,]  0.1980609811
#> [36,]  1.0000000000
#> 
#> $p
#>               [,1]       [,2]       [,3]         [,4]        [,5]         [,6]
#>  [1,] 1.000000e+00 0.11791621 0.30008979 2.757462e-05 0.239279758 0.4989690545
#>  [2,] 1.179162e-01 1.00000000 0.05439042 6.148622e-01 0.383833668 0.5510587312
#>  [3,] 3.000898e-01 0.05439042 1.00000000 9.610554e-01 0.191876788 0.7532501350
#>  [4,] 2.757462e-05 0.61486215 0.96105539 1.000000e+00 0.230108363 0.0656148662
#>  [5,] 2.392798e-01 0.38383367 0.19187679 2.301084e-01 1.000000000 0.0063902641
#>  [6,] 4.989691e-01 0.55105873 0.75325013 6.561487e-02 0.006390264 1.0000000000
#>  [7,] 9.697799e-01 0.72189902 0.79615610 7.934092e-01 0.537064811 0.0001097301
#>  [8,] 7.845785e-01 0.34056637 0.43782264 7.292092e-01 0.758737692 0.6125332791
#>  [9,] 9.509579e-01 0.78792996 0.82189557 9.346016e-01 0.816884732 0.8171525062
#> [10,] 9.260952e-01 0.88257369 0.92805260 8.030470e-01 0.883213370 0.5809854103
#> [11,] 8.064694e-01 0.22498554 0.17586973 3.883486e-01 0.431319481 0.7368828528
#> [12,] 9.584833e-01 0.68036514 0.49251366 7.910801e-01 0.444487607 0.5354783892
#> [13,] 9.602674e-01 0.92567634 0.79320255 9.067943e-01 0.488609153 0.2986655835
#> [14,] 4.228728e-01 0.04741965 0.06033388 5.775796e-01 0.361045711 0.6797817692
#> [15,] 7.400854e-01 0.42583465 0.39633771 9.503428e-01 0.500854339 0.6957151796
#> [16,] 9.790926e-01 0.94279342 0.80392521 9.285621e-01 0.617686598 0.3433462330
#> [17,] 7.837705e-01 0.86237217 0.90829882 9.965598e-01 0.915242278 0.5252431920
#> [18,] 1.472551e-01 0.60123820 0.92554327 8.048859e-02 0.735949321 0.4231779698
#> [19,] 8.772793e-01 0.82431753 0.97540547 7.805190e-01 0.863455947 0.4646043684
#> [20,] 8.012805e-01 0.90362734 0.94948169 8.008956e-01 0.992631629 0.5828804866
#> [21,] 8.157756e-01 0.85718050 0.96757101 8.468305e-01 0.991135555 0.5729398349
#> [22,] 8.383923e-01 0.92785367 0.77140820 4.678936e-01 0.878376130 0.8084816835
#> [23,] 9.327523e-01 0.92532555 0.93411822 8.368500e-01 0.669280307 0.3508752907
#> [24,] 7.416077e-01 0.74388751 0.84009318 7.872479e-01 0.901360825 0.4017040248
#> [25,] 6.814463e-01 0.99589566 0.86041940 3.654125e-01 0.722809015 0.2748164598
#> [26,] 9.745097e-01 0.77629961 0.64804587 5.222815e-01 0.815664299 0.9306071637
#> [27,] 8.200722e-01 0.89699290 0.81865559 8.834706e-01 0.899672484 0.7654177523
#> [28,] 5.817360e-01 0.93648910 0.94658044 5.644312e-01 0.825623117 0.3714183853
#> [29,] 4.918757e-01 0.28693612 0.63113608 8.918465e-01 0.624152247 0.1989793803
#> [30,] 6.759692e-01 0.71672265 0.66964438 9.360609e-01 0.814896003 0.7676159340
#> [31,] 8.457260e-01 0.94688454 0.82561105 9.370616e-01 0.842634224 0.6142857962
#> [32,] 6.267001e-01 0.85587088 0.90297286 3.087926e-01 0.627085070 0.9437441104
#> [33,] 4.735760e-01 0.92732769 0.72710401 2.133475e-01 0.494563798 0.9231060502
#> [34,] 7.706809e-01 0.93405498 0.90635928 7.081233e-01 0.995002550 0.5261461009
#> [35,] 8.748085e-01 0.84474939 0.99582010 6.377950e-01 0.676067506 0.5022085903
#> [36,] 9.748557e-01 0.96102840 0.80560969 9.256755e-01 0.597557264 0.6078369108
#>               [,7]      [,8]      [,9]     [,10]     [,11]      [,12]
#>  [1,] 0.9697799280 0.7845785 0.9509579 0.9260952 0.8064694 0.95848334
#>  [2,] 0.7218990231 0.3405664 0.7879300 0.8825737 0.2249855 0.68036514
#>  [3,] 0.7961561023 0.4378226 0.8218956 0.9280526 0.1758697 0.49251366
#>  [4,] 0.7934092449 0.7292092 0.9346016 0.8030470 0.3883486 0.79108006
#>  [5,] 0.5370648107 0.7587377 0.8168847 0.8832134 0.4313195 0.44448761
#>  [6,] 0.0001097301 0.6125333 0.8171525 0.5809854 0.7368829 0.53547839
#>  [7,] 1.0000000000 0.5426910 0.7680972 0.8341439 0.4028561 0.62435200
#>  [8,] 0.5426909972 1.0000000 0.4743625 0.6929527 0.3424783 0.64090160
#>  [9,] 0.7680972160 0.4743625 1.0000000 0.3259525 0.8548225 0.61689163
#> [10,] 0.8341438568 0.6929527 0.3259525 1.0000000 0.8464514 0.64471112
#> [11,] 0.4028560578 0.3424783 0.8548225 0.8464514 1.0000000 0.22261589
#> [12,] 0.6243520008 0.6409016 0.6168916 0.6447111 0.2226159 1.00000000
#> [13,] 0.6906132850 0.8261165 0.5154637 0.4079833 0.5301899 0.20709610
#> [14,] 0.5599037578 0.2827344 0.7536143 0.9213837 0.1047175 0.50608984
#> [15,] 0.8144286634 0.5760544 0.4809253 0.4568747 0.6445999 0.53200398
#> [16,] 0.9905009184 0.7968187 0.3730457 0.2204808 0.9184320 0.55648399
#> [17,] 0.9615588454 0.9094960 0.3362249 0.1605592 0.6592416 0.65050115
#> [18,] 0.8718569088 0.4712666 0.8888466 0.9037346 0.8136075 0.77119822
#> [19,] 0.6019107247 0.4303128 0.6040633 0.7075813 0.8017826 0.93108725
#> [20,] 0.4619441202 0.7336471 0.9213989 0.7002122 0.7628656 0.86970426
#> [21,] 0.4226200516 0.7075627 0.9385155 0.6990295 0.7682280 0.87648041
#> [22,] 0.8398904458 0.6205688 0.6990546 0.7747385 0.8774163 0.99922774
#> [23,] 0.3282881429 0.4750055 0.5858077 0.4709081 0.4148797 0.84733161
#> [24,] 0.5194933646 0.8236634 0.8500252 0.7675660 0.9794959 0.97873564
#> [25,] 0.4384645447 0.8086173 0.8595556 0.6451205 0.5439548 0.92400490
#> [26,] 0.9129030792 0.7263490 0.6850331 0.7002119 0.9977090 0.99275180
#> [27,] 0.7442963038 0.6149829 0.5872474 0.6015464 0.8810520 0.91551236
#> [28,] 0.3702648995 0.9537526 0.6933255 0.5497920 0.6201218 0.85427076
#> [29,] 0.2061067387 0.8442523 0.9042240 0.6540733 0.1115528 0.03222693
#> [30,] 0.7803131803 0.9173331 0.6343988 0.5486576 0.6964950 0.91414931
#> [31,] 0.6715306247 0.6423220 0.6002609 0.6081187 0.9522550 0.95746512
#> [32,] 0.3644820829 0.8925060 0.9679500 0.9113091 0.6405438 0.79733128
#> [33,] 0.4953381223 0.7935233 0.8930534 0.9964376 0.6373490 0.79280394
#> [34,] 0.8803796734 0.7911345 0.8094804 0.5020925 0.3003075 0.37406354
#> [35,] 0.8836252073 0.7597561 0.6184319 0.5233864 0.9764771 0.87059728
#> [36,] 0.9769687699 0.7201298 0.5998502 0.5069077 0.9701957 0.90384783
#>            [,13]      [,14]     [,15]     [,16]     [,17]      [,18]     [,19]
#>  [1,] 0.96026739 0.42287285 0.7400854 0.9790926 0.7837705 0.14725514 0.8772793
#>  [2,] 0.92567634 0.04741965 0.4258346 0.9427934 0.8623722 0.60123820 0.8243175
#>  [3,] 0.79320255 0.06033388 0.3963377 0.8039252 0.9082988 0.92554327 0.9754055
#>  [4,] 0.90679427 0.57757963 0.9503428 0.9285621 0.9965598 0.08048859 0.7805190
#>  [5,] 0.48860915 0.36104571 0.5008543 0.6176866 0.9152423 0.73594932 0.8634559
#>  [6,] 0.29866558 0.67978177 0.6957152 0.3433462 0.5252432 0.42317797 0.4646044
#>  [7,] 0.69061328 0.55990376 0.8144287 0.9905009 0.9615588 0.87185691 0.6019107
#>  [8,] 0.82611650 0.28273436 0.5760544 0.7968187 0.9094960 0.47126657 0.4303128
#>  [9,] 0.51546367 0.75361434 0.4809253 0.3730457 0.3362249 0.88884663 0.6040633
#> [10,] 0.40798332 0.92138368 0.4568747 0.2204808 0.1605592 0.90373461 0.7075813
#> [11,] 0.53018989 0.10471747 0.6445999 0.9184320 0.6592416 0.81360751 0.8017826
#> [12,] 0.20709610 0.50608984 0.5320040 0.5564840 0.6505011 0.77119822 0.9310872
#> [13,] 1.00000000 0.89787654 0.5273278 0.3667455 0.3916009 0.63117564 0.8593768
#> [14,] 0.89787654 1.00000000 0.3769518 0.8677792 0.8483932 0.88115882 0.8573586
#> [15,] 0.52732780 0.37695183 1.0000000 0.3369345 0.3487064 0.77405473 0.8561074
#> [16,] 0.36674547 0.86777916 0.3369345 1.0000000 0.1347878 0.57579359 0.9021522
#> [17,] 0.39160089 0.84839324 0.3487064 0.1347878 1.0000000 0.56310936 0.8112425
#> [18,] 0.63117564 0.88115882 0.7740547 0.5757936 0.5631094 1.00000000 0.3760501
#> [19,] 0.85937680 0.85735860 0.8561074 0.9021522 0.8112425 0.37605012 1.0000000
#> [20,] 0.67887268 0.94961907 0.7679399 0.6091600 0.5923072 0.20909535 0.5917038
#> [21,] 0.67890962 0.95772692 0.7722125 0.6030106 0.5920938 0.21957754 0.5912534
#> [22,] 0.96779158 0.87505677 0.9813510 0.9562657 0.8293415 0.46004870 0.3064226
#> [23,] 0.78084204 0.75426091 0.8642443 0.9747692 0.7429385 0.50282865 0.9933304
#> [24,] 0.90381371 0.93926601 0.9588205 0.9288063 0.9640039 0.81552145 0.7528256
#> [25,] 0.84204775 0.73264317 0.8779111 0.7057732 0.6239756 0.82347580 0.6998492
#> [26,] 0.97390884 0.74677531 0.9863243 0.8692142 0.7034193 0.59108807 0.3896967
#> [27,] 0.91580142 0.89245715 0.8134345 0.7138773 0.5555921 0.89458955 0.3076254
#> [28,] 0.65304316 0.90233708 0.6194673 0.4585799 0.3739554 0.09489470 0.8765320
#> [29,] 0.01362212 0.39660697 0.5706930 0.9022564 0.9110834 0.92264150 0.7567006
#> [30,] 0.97069960 0.75616885 0.7138158 0.5365717 0.3138619 0.47259046 0.3424451
#> [31,] 0.98590347 0.95149457 0.7715680 0.6917614 0.5207267 0.98644109 0.2896759
#> [32,] 0.95582987 0.90953474 0.8844777 0.8240935 0.9778676 0.62625146 0.9878955
#> [33,] 0.91231087 0.98057334 0.8191571 0.8168924 0.9477196 0.42985944 0.8991283
#> [34,] 0.58239555 0.67467171 0.9989822 0.7785739 0.7328420 0.16297901 0.7216697
#> [35,] 0.83210807 0.80466702 0.7113459 0.6104731 0.4188556 0.55676681 0.7003816
#> [36,] 0.85574946 0.99529104 0.8091402 0.6734988 0.4773199 0.23032923 0.5770067
#>           [,20]     [,21]     [,22]       [,23]       [,24]     [,25]     [,26]
#>  [1,] 0.8012805 0.8157756 0.8383923 0.932752298 0.741607689 0.6814463 0.9745097
#>  [2,] 0.9036273 0.8571805 0.9278537 0.925325547 0.743887506 0.9958957 0.7762996
#>  [3,] 0.9494817 0.9675710 0.7714082 0.934118224 0.840093183 0.8604194 0.6480459
#>  [4,] 0.8008956 0.8468305 0.4678936 0.836849996 0.787247885 0.3654125 0.5222815
#>  [5,] 0.9926316 0.9911356 0.8783761 0.669280307 0.901360825 0.7228090 0.8156643
#>  [6,] 0.5828805 0.5729398 0.8084817 0.350875291 0.401704025 0.2748165 0.9306072
#>  [7,] 0.4619441 0.4226201 0.8398904 0.328288143 0.519493365 0.4384645 0.9129031
#>  [8,] 0.7336471 0.7075627 0.6205688 0.475005495 0.823663410 0.8086173 0.7263490
#>  [9,] 0.9213989 0.9385155 0.6990546 0.585807697 0.850025162 0.8595556 0.6850331
#> [10,] 0.7002122 0.6990295 0.7747385 0.470908117 0.767566005 0.6451205 0.7002119
#> [11,] 0.7628656 0.7682280 0.8774163 0.414879674 0.979495943 0.5439548 0.9977090
#> [12,] 0.8697043 0.8764804 0.9992277 0.847331613 0.978735636 0.9240049 0.9927518
#> [13,] 0.6788727 0.6789096 0.9677916 0.780842045 0.903813710 0.8420477 0.9739088
#> [14,] 0.9496191 0.9577269 0.8750568 0.754260915 0.939266014 0.7326432 0.7467753
#> [15,] 0.7679399 0.7722125 0.9813510 0.864244257 0.958820528 0.8779111 0.9863243
#> [16,] 0.6091600 0.6030106 0.9562657 0.974769152 0.928806271 0.7057732 0.8692142
#> [17,] 0.5923072 0.5920938 0.8293415 0.742938544 0.964003865 0.6239756 0.7034193
#> [18,] 0.2090953 0.2195775 0.4600487 0.502828654 0.815521448 0.8234758 0.5910881
#> [19,] 0.5917038 0.5912534 0.3064226 0.993330426 0.752825573 0.6998492 0.3896967
#> [20,] 1.0000000 0.2765406 0.5507777 0.474961987 0.572837124 0.3748240 0.5810092
#> [21,] 0.2765406 1.0000000 0.5480738 0.500636510 0.619207939 0.3853316 0.5730311
#> [22,] 0.5507777 0.5480738 1.0000000 0.927202761 0.615157887 0.6662740 0.3268977
#> [23,] 0.4749620 0.5006365 0.9272028 1.000000000 0.009008753 0.1653541 0.9859362
#> [24,] 0.5728371 0.6192079 0.6151579 0.009008753 1.000000000 0.2117989 0.5933166
#> [25,] 0.3748240 0.3853316 0.6662740 0.165354053 0.211798890 1.0000000 0.6816288
#> [26,] 0.5810092 0.5730311 0.3268977 0.985936244 0.593316631 0.6816288 1.0000000
#> [27,] 0.7909045 0.7762421 0.3055968 0.924995896 0.604916175 0.7390036 0.3222152
#> [28,] 0.3104232 0.3148241 0.7876623 0.858527949 0.763611945 0.4052307 0.7815130
#> [29,] 0.7687633 0.7188534 0.9404552 0.923812309 0.928767913 0.8952911 0.9593339
#> [30,] 0.9959657 0.9781676 0.3279924 0.218476717 0.255555386 0.6967127 0.3270821
#> [31,] 0.8216151 0.8256415 0.3284575 0.563525899 0.465086433 0.6871214 0.3654866
#> [32,] 0.5183993 0.4938395 0.8353393 0.764781713 0.659689465 0.5429311 0.6882723
#> [33,] 0.4834633 0.4670724 0.7845819 0.461297506 0.772621207 0.5490297 0.6525675
#> [34,] 0.8277660 0.7989126 0.7333919 0.114488791 0.300499603 0.4390265 0.5323916
#> [35,] 0.9650956 0.9077933 0.8167366 0.788524667 0.702029605 0.9921652 0.7708019
#> [36,] 0.7695424 0.7865755 0.6357800 0.957609121 0.881847245 0.9525981 0.5919377
#>           [,27]     [,28]       [,29]     [,30]     [,31]      [,32]      [,33]
#>  [1,] 0.8200722 0.5817360 0.491875692 0.6759692 0.8457260 0.62670007 0.47357599
#>  [2,] 0.8969929 0.9364891 0.286936117 0.7167226 0.9468845 0.85587088 0.92732769
#>  [3,] 0.8186556 0.9465804 0.631136076 0.6696444 0.8256111 0.90297286 0.72710401
#>  [4,] 0.8834706 0.5644312 0.891846477 0.9360609 0.9370616 0.30879260 0.21334753
#>  [5,] 0.8996725 0.8256231 0.624152247 0.8148960 0.8426342 0.62708507 0.49456380
#>  [6,] 0.7654178 0.3714184 0.198979380 0.7676159 0.6142858 0.94374411 0.92310605
#>  [7,] 0.7442963 0.3702649 0.206106739 0.7803132 0.6715306 0.36448208 0.49533812
#>  [8,] 0.6149829 0.9537526 0.844252316 0.9173331 0.6423220 0.89250598 0.79352330
#>  [9,] 0.5872474 0.6933255 0.904224042 0.6343988 0.6002609 0.96795002 0.89305339
#> [10,] 0.6015464 0.5497920 0.654073300 0.5486576 0.6081187 0.91130912 0.99643760
#> [11,] 0.8810520 0.6201218 0.111552827 0.6964950 0.9522550 0.64054377 0.63734901
#> [12,] 0.9155124 0.8542708 0.032226926 0.9141493 0.9574651 0.79733128 0.79280394
#> [13,] 0.9158014 0.6530432 0.013622115 0.9706996 0.9859035 0.95582987 0.91231087
#> [14,] 0.8924571 0.9023371 0.396606969 0.7561689 0.9514946 0.90953474 0.98057334
#> [15,] 0.8134345 0.6194673 0.570693030 0.7138158 0.7715680 0.88447767 0.81915715
#> [16,] 0.7138773 0.4585799 0.902256412 0.5365717 0.6917614 0.82409352 0.81689239
#> [17,] 0.5555921 0.3739554 0.911083449 0.3138619 0.5207267 0.97786762 0.94771958
#> [18,] 0.8945896 0.0948947 0.922641504 0.4725905 0.9864411 0.62625146 0.42985944
#> [19,] 0.3076254 0.8765320 0.756700631 0.3424451 0.2896759 0.98789549 0.89912832
#> [20,] 0.7909045 0.3104232 0.768763323 0.9959657 0.8216151 0.51839932 0.48346335
#> [21,] 0.7762421 0.3148241 0.718853394 0.9781676 0.8256415 0.49383951 0.46707243
#> [22,] 0.3055968 0.7876623 0.940455180 0.3279924 0.3284575 0.83533930 0.78458188
#> [23,] 0.9249959 0.8585279 0.923812309 0.2184767 0.5635259 0.76478171 0.46129751
#> [24,] 0.6049162 0.7636119 0.928767913 0.2555554 0.4650864 0.65968947 0.77262121
#> [25,] 0.7390036 0.4052307 0.895291108 0.6967127 0.6871214 0.54293114 0.54902973
#> [26,] 0.3222152 0.7815130 0.959333942 0.3270821 0.3654866 0.68827227 0.65256753
#> [27,] 1.0000000 0.7871195 0.997952086 0.1644570 0.2267954 0.95819641 0.95399553
#> [28,] 0.7871195 1.0000000 0.920998361 0.5308674 0.7586729 0.43303755 0.39703588
#> [29,] 0.9979521 0.9209984 1.000000000 0.5886286 0.7504086 0.41620283 0.29016489
#> [30,] 0.1644570 0.5308674 0.588628625 1.0000000 0.1227200 0.89381154 0.79964537
#> [31,] 0.2267954 0.7586729 0.750408636 0.1227200 1.0000000 0.95619145 0.95864284
#> [32,] 0.9581964 0.4330376 0.416202833 0.8938115 0.9561914 1.00000000 0.05084406
#> [33,] 0.9539955 0.3970359 0.290164890 0.7996454 0.9586428 0.05084406 1.00000000
#> [34,] 0.6111426 0.4967509 0.007534501 0.5985690 0.9392609 0.04514332 0.13583771
#> [35,] 0.7636102 0.9025781 0.440789975 0.6228927 0.6587794 0.41016159 0.43819449
#> [36,] 0.7276361 0.6285420 0.657626148 0.7291570 0.6621477 0.40805303 0.35820322
#>             [,34]      [,35]      [,36]
#>  [1,] 0.770680917 0.87480852 0.97485575
#>  [2,] 0.934054982 0.84474939 0.96102840
#>  [3,] 0.906359285 0.99582010 0.80560969
#>  [4,] 0.708123322 0.63779501 0.92567548
#>  [5,] 0.995002550 0.67606751 0.59755726
#>  [6,] 0.526146101 0.50220859 0.60783691
#>  [7,] 0.880379673 0.88362521 0.97696877
#>  [8,] 0.791134485 0.75975608 0.72012984
#>  [9,] 0.809480420 0.61843190 0.59985021
#> [10,] 0.502092493 0.52338637 0.50690771
#> [11,] 0.300307489 0.97647707 0.97019572
#> [12,] 0.374063536 0.87059728 0.90384783
#> [13,] 0.582395548 0.83210807 0.85574946
#> [14,] 0.674671709 0.80466702 0.99529104
#> [15,] 0.998982181 0.71134588 0.80914018
#> [16,] 0.778573938 0.61047311 0.67349882
#> [17,] 0.732841966 0.41885556 0.47731995
#> [18,] 0.162979015 0.55676681 0.23032923
#> [19,] 0.721669666 0.70038164 0.57700670
#> [20,] 0.827766041 0.96509555 0.76954240
#> [21,] 0.798912571 0.90779334 0.78657548
#> [22,] 0.733391913 0.81673662 0.63577995
#> [23,] 0.114488791 0.78852467 0.95760912
#> [24,] 0.300499603 0.70202961 0.88184725
#> [25,] 0.439026464 0.99216524 0.95259813
#> [26,] 0.532391575 0.77080189 0.59193773
#> [27,] 0.611142612 0.76361021 0.72763613
#> [28,] 0.496750869 0.90257812 0.62854202
#> [29,] 0.007534501 0.44078998 0.65762615
#> [30,] 0.598568999 0.62289270 0.72915703
#> [31,] 0.939260867 0.65877944 0.66214769
#> [32,] 0.045143319 0.41016159 0.40805303
#> [33,] 0.135837711 0.43819449 0.35820322
#> [34,] 1.000000000 0.84854078 0.83523502
#> [35,] 0.848540780 1.00000000 0.06754662
#> [36,] 0.835235018 0.06754662 1.00000000
#>