Create an adjacency matrix based on random forest

`randomForest` infers an adjacency matrix using random forest using the `GENIE3` function from the `GENIE3` package. `randomForest` returns the importance of the link between features in the form of an adjacency matrix.

randomForest(x, ...)

Arguments

x

matrix, where columns are the samples and the rows are features (metabolites), cell entries are intensity values

...

parameters passed to `GENIE3`

Value

matrix, matrix with the importance of the links inferred from random forest algorithm implemented by `GENIE3`

Details

For use of the parameters used in the `GENIE3` function, refer to `?GENIE3::GENIE3`. The arguments `regulators` and `targets` are set to `NULL`. Element \(w_{i,j}\) (row i, column j) gives the importance of the link from i to j.

Author

Thomas Naake, thomasnaake@googlemail.com

Examples

data("x_test", package = "MetNet")
x <- x_test[1:10, 3:ncol(x_test)]
x <- as.matrix(x)
randomForest(x)
#>             x9485      x7449     x11179     x11374     x13007      x14826
#> x9485  0.00000000 0.15973554 0.10082017 0.47069825 0.14877643 0.090452641
#> x7449  0.14891683 0.00000000 0.26442881 0.06984343 0.06725951 0.009854983
#> x11179 0.10143209 0.31521071 0.00000000 0.06721923 0.34655768 0.151581713
#> x11374 0.44395573 0.04050264 0.01528082 0.00000000 0.08684377 0.117916817
#> x13007 0.17481283 0.14237170 0.33156723 0.19179020 0.00000000 0.420871246
#> x14826 0.06256378 0.01376658 0.03766922 0.09244387 0.25051612 0.000000000
#> x16257 0.01096601 0.01263053 0.00926449 0.03482874 0.01963171 0.139854463
#> x10804 0.02309817 0.15064430 0.11340186 0.01656644 0.02837165 0.008539627
#> x12669 0.01991379 0.10466464 0.08992923 0.02954326 0.03138255 0.029141842
#> x14553 0.01434077 0.06047337 0.03763818 0.02706658 0.02066057 0.031786667
#>            x16257      x10804     x12669     x14553
#> x9485  0.03183114 0.032250528 0.02949393 0.02328735
#> x7449  0.01870877 0.115767972 0.06673537 0.02393510
#> x11179 0.04366904 0.153902040 0.12205795 0.08275852
#> x11374 0.08962916 0.016440269 0.03209368 0.07511988
#> x13007 0.16378928 0.030583019 0.02418664 0.01965548
#> x14826 0.29822483 0.008279003 0.01142915 0.02506818
#> x16257 0.00000000 0.013923006 0.01725544 0.04968118
#> x10804 0.06491700 0.000000000 0.26870497 0.21984227
#> x12669 0.12610653 0.405616352 0.00000000 0.48065205
#> x14553 0.16312426 0.223237810 0.42804288 0.00000000