Return consensus ranks from a matrix containing ranks

`topKnet` returns consensus ranks depending on the `type` argument from `ranks`, a matrix containing the ranks per statistical `model`.

topKnet(ranks, type, na.rm = TRUE)

Arguments

ranks

`matrix` containing the ranks per statistical model (in columns) and per feature pair (in rows)

type

`character`, either `"top1"`, `"top2"` or `"mean"`

na.rm

`logical`, if set to `TRUE`, the `NA`s in the assay slots will not be taken into account when creating the `"top1"`, `"top2"` or `"mean"` of ranks. If set to `FALSE`, the `NA`s will be taken into account when creating the `"top1"`, `"top2"` or `"mean"` ranks. If `FALSE` the resulting aggregations will be `NA` if an `NA` is present in the coeffients of one feature pair.

Value

`numeric` `vector“ with consensus ranks

Details

See Hase et al. (2014) for further details.

References

Hase et al. (2014): Harnessing Diversity towards the Reconstructing of Large Scale Gene Regulatory Networks. PLoS Computational Biology, 2013, e1003361, doi: [10.1371/journal.pcbi.1003361](https://journals.plos.org/ploscompbiol/article?id=10.1371/journal.pcbi.1003361)

Author

Thomas Naake, thomasnaake@googlemail.com

Examples

## na.rm == TRUE
ranks <- matrix(c(c(1, 2, 3), c(2, 1, 3)), ncol = 2)

## type = "top1"
MetNet:::topKnet(ranks = ranks, type = "top1", na.rm = TRUE)
#> [1] 1 1 3

## type = "top2"
MetNet:::topKnet(ranks = ranks, type = "top2", na.rm = TRUE)
#> [1] 2 2 3

## type = "mean"
MetNet:::topKnet(ranks = ranks, type = "mean", na.rm = TRUE)
#> [1] 1.5 1.5 3.0

## na.rm == FALSE
ranks <- matrix(c(c(1, 2, 3), c(2, 1, 3)), ncol = 2)

## type = "top1"
MetNet:::topKnet(ranks = ranks, type = "top1", na.rm = FALSE)
#> [1] 1 1 3

## type = "top2"
MetNet:::topKnet(ranks = ranks, type = "top2", na.rm = FALSE)
#> [1] 2 2 3

## type = "mean"
MetNet:::topKnet(ranks = ranks, type = "mean", na.rm = FALSE)
#> [1] 1.5 1.5 3.0

## na.rm == FALSE
ranks <- matrix(c(c(1, 2, NA), c(2, 1, 3)), ncol = 2)

## type = "top1"
MetNet:::topKnet(ranks = ranks, type = "top1", na.rm = FALSE)
#> [1]  1  1 NA

## type = "top2"
MetNet:::topKnet(ranks = ranks, type = "top2", na.rm = FALSE)
#> [1]  2  2 NA

## type = "mean"
MetNet:::topKnet(ranks = ranks, type = "mean", na.rm = FALSE)
#> [1] 1.5 1.5  NA