Create an adjacency matrix based on score-based structure learning algorithm

`bayes` infers an adjacency matrix using score-based structure learning algorithm `boot.strength` from the `bnlearn` package. `bayes` extracts then the reported connections from running the `boot.strength` function and assigns the strengths of the arcs of the Bayesian connections to an adjacency matrix. `bayes` returns this weighted adjacency matrix.

bayes(x, algorithm = "tabu", R = 100, ...)

Arguments

x
`matrix` where columns are the samples and the rows are features (metabolites), cell entries are intensity values

algorithm
`character`, structure learning to be applied to the bootstrap replicates (default is `"tabu"`)

R
`numeric`, number of bootstrap replicates

...
parameters passed to `boot.strength`

Value

`matrix` with edges inferred from score-based structure learning algorithm `boot.strength`

Details

`boot.strength` measures the strength of the probabilistic relationships by the arcs of a Bayesian network, as learned from bootstrapped data. By default `bayes` uses the Tabu greedy search.

For use of the parameters used in the `boot.strength` function, refer to `?bnlearn::boot.strength`. For further information see also Friedman et al. (1999) and Scutari and Nagarajan (2001).

References

Friedman et al. (1999): Data Analysis with Bayesian Networks: A Bootstrap Approach. Proceedings of the 15th Annual Conference on Uncertainty in Artificial Intelligence, 196-201.

Scutari and Nagarajan (2011): On Identifying Significant Edges in Graphical Models. Proceedings of the Workshop Probabilistic Problem Solving in Biomedicine of the 13th Artificial Intelligence in Medicine Conference, 15-27.

Author

Thomas Naake, thomasnaake@googlemail.com

Examples

data("x_test", package = "MetNet")
x <- x_test[1:10, 3:ncol(x_test)]
x <- as.matrix(x)
bayes(x, algorithm = "tabu", R = 100)

#>        x9485 x7449 x11179 x11374 x13007 x14826 x16257 x10804 x12669 x14553
#> x9485   0.00  0.96   0.83   1.00   0.95   0.18   0.25   0.30   0.36   0.07
#> x7449   0.96  0.00   1.00   0.61   0.69   0.22   0.37   0.76   0.22   0.11
#> x11179  0.83  1.00   0.00   0.93   1.00   0.97   0.14   0.41   0.15   0.44
#> x11374  1.00  0.61   0.93   0.00   0.34   0.81   0.64   0.11   0.14   0.51
#> x13007  0.95  0.69   1.00   0.34   0.00   1.00   0.99   0.06   0.29   0.43
#> x14826  0.18  0.22   0.97   0.81   1.00   0.00   1.00   0.07   0.29   0.65
#> x16257  0.25  0.37   0.14   0.64   0.99   1.00   0.00   0.14   0.29   0.93
#> x10804  0.30  0.76   0.41   0.11   0.06   0.07   0.14   0.00   1.00   1.00
#> x12669  0.36  0.22   0.15   0.14   0.29   0.29   0.29   1.00   0.00   1.00
#> x14553  0.07  0.11   0.44   0.51   0.43   0.65   0.93   1.00   1.00   0.00