Create an adjacency matrix based on correlation

`correlation` infers an adjacency matrix using correlation using the `cor` function (from the `stats` package), `pcor` (from `ppcor`) or `spcor` (from `ppcor`). `correlation` extracts the reported pair-wise correlation coefficients from the function `corAndPvalue`, `pcor` or `spcor` and will return the weighted adjacency matrix of the absolute correlation values.

correlation(x, type = "pearson", use = "pairwise.complete.obs")

Arguments

x
matrix, where columns are the samples and the rows are features (metabolites), cell entries are intensity values

type
`character`, either "pearson", "spearman", "pearson_partial", "spearman_partial", "pearson_semipartial" or "spearman_semipartial".

use
`character` string giving a method for computing covariance in the presence of missing values, Only for `type = "pearson"` or `type = "spearman"`. For further information see `?stats::cor`

Value

matrix, matrix with edges inferred from correlation algorithm `corAndPvalue`, `pcor` or `spcor` (depending on the chosen `method`)

Details

If `"pearson"` or `"spearman"` is used as a `method`, the function `corAndPvalue` from `stats` will be employed.

If `"pearson_partial"` or `"spearman_partial"?` is used as a `method` the function `pcor` from `spcor` will be employed.

If `"pearson_semipartial"` or `"spearman_semipartial"` is used as a `method` the function `spcor` from `spcor` will be employed.

`type` will be passed to argument `method` in `cor` (in the case of `"pearson"` or `"spearman"`) or to `method` in `pcor` (`"pearson"` and `"spearman"` for `"pearson_partial"` and `"spearman_partial"`, respectively) or to `method` in `spcor` (`"pearson"` or `"spearman"` for `"pearson_semipartial"` and `"spearman_semipartial"`, respectively).

Author

Thomas Naake, thomasnaake@googlemail.com

Examples

data("x_test", package = "MetNet")
x <- x_test[1:10, 3:ncol(x_test)]
x <- as.matrix(x)
correlation(x, type = "pearson")

#>            x9485     x7449     x11179    x11374    x13007    x14826     x16257
#> x9485  1.0000000 0.8000351 0.78391122 0.9403697 0.8510633 0.6986100 0.27631528
#> x7449  0.8000351 1.0000000 0.92406116 0.6044475 0.7470669 0.4622945 0.01600220
#> x11179 0.7839112 0.9240612 1.00000000 0.5948374 0.8648860 0.5646757 0.01969613
#> x11374 0.9403697 0.6044475 0.59483740 1.0000000 0.7782440 0.7367013 0.35634911
#> x13007 0.8510633 0.7470669 0.86488604 0.7782440 1.0000000 0.8600956 0.32016094
#> x14826 0.6986100 0.4622945 0.56467568 0.7367013 0.8600956 1.0000000 0.68398191
#> x16257 0.2763153 0.0160022 0.01969613 0.3563491 0.3201609 0.6839819 1.00000000
#> x10804 0.5578105 0.7620814 0.74124973 0.4150268 0.6205390 0.3859232 0.09495813
#> x12669 0.5059605 0.6519108 0.66648908 0.4345848 0.6217924 0.4400734 0.12757798
#> x14553 0.3993292 0.5008692 0.54831042 0.3736941 0.5383033 0.3926268 0.19360513
#>            x10804    x12669    x14553
#> x9485  0.55781050 0.5059605 0.3993292
#> x7449  0.76208143 0.6519108 0.5008692
#> x11179 0.74124973 0.6664891 0.5483104
#> x11374 0.41502679 0.4345848 0.3736941
#> x13007 0.62053902 0.6217924 0.5383033
#> x14826 0.38592321 0.4400734 0.3926268
#> x16257 0.09495813 0.1275780 0.1936051
#> x10804 1.00000000 0.8902988 0.7161565
#> x12669 0.89029884 1.0000000 0.9398149
#> x14553 0.71615645 0.9398149 1.0000000