Create an adjacency matrix based on random forest

`randomForest` infers an adjacency matrix using random forest using the `GENIE3` function from the `GENIE3` package. `randomForest` returns the importance of the link between features in the form of an adjacency matrix.

randomForest(x, ...)

Arguments

x
matrix, where columns are the samples and the rows are features (metabolites), cell entries are intensity values

...
parameters passed to `GENIE3`

Value

matrix, matrix with the importance of the links inferred from random forest algorithm implemented by `GENIE3`

Details

For use of the parameters used in the `GENIE3` function, refer to `?GENIE3::GENIE3`. The arguments `regulators` and `targets` are set to `NULL`. Element \(w_{i,j}\) (row i, column j) gives the importance of the link from i to j.

Author

Thomas Naake, thomasnaake@googlemail.com

Examples

data("x_test", package = "MetNet")
x <- x_test[1:10, 3:ncol(x_test)]
x <- as.matrix(x)
randomForest(x)

#>            x10804     x11179     x11374     x12669     x13007     x14553
#> x10804 0.00000000 0.10724091 0.01744763 0.28201701 0.02619574 0.22371318
#> x11179 0.15215007 0.00000000 0.07446625 0.14213103 0.34533540 0.07958550
#> x11374 0.01603156 0.01616277 0.00000000 0.03008362 0.09642533 0.06691177
#> x12669 0.39704282 0.09686057 0.02794283 0.00000000 0.03351563 0.47503888
#> x13007 0.03268908 0.34467228 0.17499583 0.02445993 0.00000000 0.01932231
#> x14553 0.21537626 0.03117395 0.02850617 0.39345255 0.02239549 0.00000000
#> x14826 0.01102858 0.03083392 0.10135006 0.01125148 0.23411533 0.02908174
#> x16257 0.01546201 0.01041634 0.03354644 0.01769401 0.01977308 0.05399992
#> x7449  0.12884992 0.25882254 0.07343976 0.06905078 0.05913272 0.02617157
#> x9485  0.03136970 0.10381671 0.46830503 0.02985960 0.16311129 0.02617513
#>             x14826     x16257      x7449      x9485
#> x10804 0.009807396 0.06511375 0.14144841 0.02256722
#> x11179 0.160931057 0.04367634 0.29229748 0.09867864
#> x11374 0.112860602 0.08119757 0.03948059 0.43060914
#> x12669 0.026115273 0.12755216 0.12049790 0.02162894
#> x13007 0.415295824 0.17916197 0.16206164 0.17867171
#> x14553 0.029730036 0.17223621 0.05679406 0.01387617
#> x14826 0.000000000 0.28367005 0.01371166 0.06911926
#> x16257 0.138674894 0.00000000 0.01351034 0.01314165
#> x7449  0.010180891 0.01720281 0.00000000 0.15170726
#> x9485  0.096404027 0.03018915 0.16019793 0.00000000